### 2022-08-25 #4709 @MartianColonist
Here are some potential reviewers with experience in the exoplanet field (and no ongoing JOSS reviews, according to the reviewer spreadsheet): Néstor Espinoza (nespinoza), Adina Feinstein (afeinstein20), or Wilfred Tyler Gee (wtgee).
