### 2022-08-27 d2929eb
Added emission spectrum plotting
