+++ POSEIDON/visuals.py
@@ -872,9 +872,11 @@ def plot_spectra(spectra, planet, data_properties = None,
                  plot_full_res = True, bin_spectra = True, R_to_bin = 100, 
                  wl_min = None, wl_max = None, transit_depth_min = None,
                  transit_depth_max = None, show_data = False, plt_label = None,
+                 FpFs_min = None, FpFs_max = None, y_unit = 'transit_depth', 
                  colour_list = [], spectra_labels = [], data_colour_list = [],
                  data_labels = [], data_marker_list = [], 
-                 data_marker_size_list = [], figure_shape = 'default'):
+                 data_marker_size_list = [], 
+                 wl_axis = 'log', figure_shape = 'default'):
     ''' 
     Plot a collection of individual model transmission spectra.
     
@@ -973,50 +975,96 @@ def plot_spectra(spectra, planet, data_properties = None,
             wl_max = max(wl_max, wl_max_i)
 
     wl_range = wl_max - wl_min
-    
+   
     # If the user did not specify a transit depth range, find min and max from input models
-    if (transit_depth_min == None):
-        
-        transit_depth_min_plt = 1e10   # Dummy value
-        
-        # Loop over each model, finding the most extreme min / max range 
-        for i in range(N_spectra):
-            
-            transit_depth_min_i = np.min(spectra[i][0])
-            transit_depth_min_plt = min(transit_depth_min_plt, transit_depth_min_i)
-            
-        # Check if the lowest data point falls below the current y-limit
-        if (show_data == True):
-            if (transit_depth_min_plt > min(ydata - err_data)):
-                
-                transit_depth_min_plt = min(ydata - err_data)
-            
-        transit_depth_min_plt = 0.995*transit_depth_min_plt  # Extend slightly below
-        
-    else:
-        transit_depth_min_plt = transit_depth_min
+    if (y_unit in ['(Rp/Rs)^2', '(Rp/R*)^2', 'transit_depth']):
 
-    # If the user did not specify a transit depth range, find min and max from input models
-    if (transit_depth_max == None):
-        
-        transit_depth_max_plt = 1e-10  # Dummy value
-        
-        # Loop over each model, finding the most extreme min / max range 
-        for i in range(N_spectra):
+        if (transit_depth_min == None):
             
-            transit_depth_max_i = np.max(spectra[i][0])
-            transit_depth_max_plt = max(transit_depth_max_plt, transit_depth_max_i)
+            y_min_plt = 1e10   # Dummy value
             
-        # Check if the highest data point falls above the current y-limit
-        if (show_data == True):
-            if (transit_depth_max_plt < max(ydata + err_data)):
+            # Loop over each model, finding the most extreme min / max range 
+            for i in range(N_spectra):
                 
-                transit_depth_max_plt = max(ydata + err_data)
+                transit_depth_min_i = np.min(spectra[i][0])
+                y_min_plt = min(y_min_plt, transit_depth_min_i)
+                
+            # Check if the lowest data point falls below the current y-limit
+            if (show_data == True):
+                if (y_min_plt > min(ydata - err_data)):
+                    
+                    y_min_plt = min(ydata - err_data)
+                
+            y_min_plt = 0.995*y_min_plt  # Extend slightly below
             
[…]
