### 2022-09-11 608434c
Emission retrieval refinements
