### 2022-09-11 f3b0a45
Emission retrieval refinements
