+++ POSEIDON/core.py
@@ -211,7 +211,7 @@ def define_model(model_name, bulk_species, param_species,
                  PT_dim = 1, X_dim = 1, cloud_dim = 1, TwoD_type = None, 
                  TwoD_param_scheme = 'difference', species_EM_gradient = [], 
                  species_DN_gradient = [], species_vert_gradient = [],
-                 surface = False):
+                 surface = False, spectrum_type = 'emission'):
     '''
     Create the model dictionary defining the configuration of the user-specified 
     forward model or retrieval.
@@ -347,7 +347,7 @@ def define_model(model_name, bulk_species, param_species,
                                       X_dim, cloud_dim, TwoD_type, TwoD_param_scheme, 
                                       species_EM_gradient, species_DN_gradient, 
                                       species_vert_gradient, Atmosphere_dimension,
-                                      opaque_Iceberg, surface)
+                                      opaque_Iceberg, surface, spectrum_type)
 
     # Package model properties
     model = {'model_name': model_name, 'object_type': object_type,
@@ -952,6 +952,9 @@ def compute_spectrum(planet, star, model, atmosphere, opac, wl,
         # Convert planet surface flux to observed flux at Earth
         F_p_obs = (R_p / d)**2 * F_p
 
+        spectrum = F_p_obs
+
+        '''
         # For direct emission spectra (brown dwarfs and directly imaged planets)        
         if (spectrum_type == 'direct_emission'):
 
@@ -973,6 +976,7 @@ def compute_spectrum(planet, star, model, atmosphere, opac, wl,
 
             # Final spectrum is the planet-star flux ratio
             spectrum = F_p_obs / F_s_obs
+        '''
 
     # Write spectrum to file
     if (save_spectrum == True):
