### 2022-09-27 497c395
Added error bar binning from Spectres
