### 2022-09-28 5ba9825
Changed default spectrum type
