### 2022-10-04 fb14f0f
Deprecated multinode opacity memory (pending fix)
