### 2022-10-04 4fb9ff9
Added photosphere radius to emission spectra
