### 2022-10-04 1a7eb73
Plot improvements
