+++ POSEIDON/instrument.py
@@ -340,17 +340,17 @@ def make_model_data(spectrum, wl, sigma, sensitivity, bin_left, bin_cent,
         for n in range(N_bins):
             
             # Extend convolution beyond bin edge by max(1, 2 PSF std) model grid spaces (std rounded to integer)
-            extention = max(1, int(2 * sigma[n]))   
+            extension = max(1, int(2 * sigma[n]))   
             
             # Convolve spectrum with PSF width appropriate for a given bin 
-            spectrum_conv = gauss_conv(spectrum[(bin_left[n]-extention):(bin_right[n]+extention)], 
+            spectrum_conv = gauss_conv(spectrum[(bin_left[n]-extension):(bin_right[n]+extension)], 
                                        sigma=sigma[n], mode='nearest')
 
             # Catch a (surprisingly common) error
-            if (len(spectrum_conv[extention:-extention]) != len(sensitivity[bin_left[n]:bin_right[n]])):
+            if (len(spectrum_conv[extension:-extension]) != len(sensitivity[bin_left[n]:bin_right[n]])):
                 raise Exception("Error: Model wavelength range not wide enough to encompass all data.")
 
-            integrand = spectrum_conv[extention:-extention] * sensitivity[bin_left[n]:bin_right[n]]
+            integrand = spectrum_conv[extension:-extension] * sensitivity[bin_left[n]:bin_right[n]]
         
             # Integrate convolved spectrum over instrument sensitivity function
             data[n] = trapz(integrand, wl[bin_left[n]:bin_right[n]])   
