### 2022-10-08 6ef6f89
Added NRS specifications for G395H
