+++ POSEIDON/instrument.py
@@ -205,6 +205,8 @@ def init_instrument(wl, wl_data, half_width, instrument):
     elif (instrument.startswith('JWST')): 
         if (instrument == 'JWST_NIRSpec_Prism'): # Catch common misspelling of PRISM
             instrument = 'JWST_NIRSpec_PRISM'
+        if ('NRS' in instrument):                # If G395H split into detectors, use common sensitivity function
+            instrument = 'JWST_NIRSpec_G395H'
         sens_file = inst_dir + '/JWST/' + instrument + '_sensitivity.dat'
     
     # If instrument does not have a known sensitivity function, just use a top hat
