+++ POSEIDON/visuals.py
@@ -1,11 +1,13 @@
 # Plotting routines to visualise POSEIDON output*****
 
 from enum import unique
+import os
 import numpy as np
 import scipy.constants as sc
 import colorsys
 import matplotlib
 from pylab import rcParams
+import pymultinest
 import matplotlib.style
 from matplotlib import ticker
 import matplotlib.pyplot as plt
@@ -33,9 +35,11 @@ warnings.filterwarnings("ignore", message="This figure includes Axes that are " 
                                            "not compatible with tight_layout, " +
                                            "so results might be incorrect.")
 
-from .utility import bin_spectrum, generate_latex_param_names, round_sig_figs
+from .utility import bin_spectrum, generate_latex_param_names, round_sig_figs, \
+                     confidence_intervals
 from .instrument import bin_spectrum_to_data
-
+from .atmosphere import compute_O_to_H, compute_C_to_O
+from .parameters import split_params
               
 # Define some more flexible linestyles for convenience
 linestyles = {
@@ -968,8 +972,8 @@ def plot_spectra(spectra, planet, data_properties = None,
                  FpFs_min = None, FpFs_max = None, y_unit = 'transit_depth', 
                  colour_list = [], spectra_labels = [], data_colour_list = [],
                  data_labels = [], data_marker_list = [], 
-                 data_marker_size_list = [], 
-                 wl_axis = 'log', figure_shape = 'default'):
+                 data_marker_size_list = [], wl_axis = 'log', 
+                 figure_shape = 'default', legend_location = 'upper right'):
     ''' 
     Plot a collection of individual model transmission spectra.
     
@@ -979,6 +983,8 @@ def plot_spectra(spectra, planet, data_properties = None,
         plot_type = 'transmission'
     elif (y_unit in ['Fp/Fs', 'Fp/F*']):
         plot_type = 'emission'
+    elif (y_unit in ['Fp']):
+        plot_type = 'direct_emission'
     
     # Find number of spectra to plot
     N_spectra = len(spectra)
@@ -1325,6 +1331,8 @@ def plot_spectra(spectra, planet, data_properties = None,
         ax1.set_ylabel(r'Transit Depth $(R_p/R_*)^2$', fontsize = 16)
     elif (plot_type == 'emission'):
         ax1.set_ylabel(r'Emission Spectrum $(F_p/F_*)$', fontsize = 16)
+    elif (plot_type == 'direct_emission'):
+        ax1.set_ylabel(r'$F_{\rm{p}}$ (W m$^{-2}$ m$^{-1}$)', fontsize = 16)
 
     # Add planet name label
     ax1.text(0.02, 0.96, planet_name, horizontalalignment='left', 
@@ -1340,7 +1348,7 @@ def plot_spectra(spectra, planet, data_properties = None,
     # Plot wl tick labels
     ax1.set_xticks(wl_ticks)
     
-    legend = ax1.legend(loc='upper right', shadow=True, prop={'size':10}, 
+    legend = ax1.legend(loc=legend_location, shadow=True, prop={'size':10}, 
                         ncol=1, frameon=True)    #legend settings
   #  legend.set_bbox_to_anchor([0.75, 0.98], transform=None)
     frame = legend.get_frame()
@@ -1451,7 +1459,7 @@ def plot_data(data, planet, wl_min = None, wl_max = None,
             y_max_plt = transit_depth_max
 
     # If the user did not specify an Fp/Fs range, find min and max from input models
[…]
