### 2022-10-13 3027fc2
Bug fixes for direct flux retrievals
