### 2022-10-13 faee6e5
Bug fixes for direct flux retrievals
