### 2022-10-14 a793a91
Improved elemental ratio plot
