### 2022-10-15 80921a0
Initial GPU prototype for emission spectra
