+++ POSEIDON/emission.py
@@ -4,8 +4,16 @@ Radiative transfer calculations for generating emission spectra.
 '''
 
 import numpy as np
-from numba.core.decorators import jit
 import scipy.constants as sc
+from numba import jit, cuda
+import cupy as cp
+import math
+import os
+
+from .utility import interp_GPU
+
+block = int(os.environ['block'])
+thread = int(os.environ['thread'])
 
 
 @jit(nopython = True)
@@ -49,6 +57,44 @@ def planck_lambda_arr(T, wl):
             B_lambda[i,k] = coeff * (1.0 / (np.exp(c_2 / (wl_m[k] * T[i])) - 1.0))
             
     return B_lambda
+
+
+@cuda.jit
+def planck_lambda_arr_GPU(T, wl, B_lambda):
+    '''
+    GPU variant of the 'planck_lambda_arr' function.
+
+    Compute the Planck function spectral radiance for a range of model
+    wavelengths and atmospheric temperatures.
+    Args:
+        T (np.array of float):
+            Array of temperatures in each atmospheric layer (K).
+        wl (np.array of float): 
+            Wavelength grid (μm).
+    
+    Returns:
+        B_lambda (2D np.array of float):
+            Planck function spectral radiance as a function of layer temperature
+            and wavelength in SI units (W/m^2/sr/m).
+    '''
+    
+    thread = cuda.grid(1)
+    stride = cuda.gridsize(1)
+    
+    # Second radiative constant
+    c_2 = (sc.h * sc.c) / sc.k
+    
+    # For each wavelength
+    for k in range(thread, len(wl), stride):
+        
+        # Precompute Planck function coefficient prefactor
+        coeff = (2.0 * sc.h * sc.c**2) / (wl[k]**5)
+        
+        # For each atmospheric layer
+        for i in range(len(T)):
+            
+            # Evaluate Planck function spectral radiance
+            B_lambda[i,k] = coeff * (1.0 / (math.exp(c_2 / (wl[k] * T[i])) - 1.0))
 
 
 @jit(nopython = True)
@@ -122,6 +168,87 @@ def emission_rad_transfer(T, dz, wl, kappa, Gauss_quad = 2):
     return F, dtau
 
 
+def emission_rad_transfer_GPU(T, dz, wl, kappa, Gauss_quad = 2):
+    '''
+    GPU variant of the 'emission_rad_transfer' function.
+
+    Compute the emergent top-of-atmosphere flux from a planet or brown dwarf.
+
+    This function considers only pure thermal emission (i.e. no scattering).
+
+    Args:
+        T (np.array of float):
+            Temperatures in each atmospheric layer (K).
+        dz (np.array of float):
+            Vertical extent of each atmospheric layer (m).
+        wl (np.array of float): 
+            Wavelength grid (μm).
+        kappa (2D np.array of float):
+            Extinction coefficient in each layer as a function of wavelength (m^-1).
+        Gauss_quad (int):
+            Gaussian quadrature order for integration over emitting surface
+            (Options: 2 / 3).
+    
+    Returns:
+        F (np.array of float):
+            Spectral surface flux in SI units (W/m^2/sr/m).
+    '''
+    
+    # Load weights and cos(theta) values for desired Gaussian quadrature scheme
+    if (Gauss_quad == 2):
+        W = cp.array([0.5, 0.5])
+        mu = cp.array([0.5 - 0.5*cp.sqrt(1.0/3.0), 0.5 + 0.5*cp.sqrt(1.0/3.0)])
+    elif (Gauss_quad == 3):
+        W = cp.array([5.0/18.0, 4.0/9.0, 5.0/18.0])
[…]
