### 2022-10-15 c6bc532
Initial GPU prototype for emission spectra
