+++ POSEIDON/core.py
@@ -13,7 +13,11 @@ os.environ['OPENBLAS_NUM_THREADS'] = '1'
 os.environ['MKL_NUM_THREADS'] = '1'
 os.environ['CBLAS_NUM_THREADS'] = '1'
 
+os.environ['block'] = '128'
+os.environ['thread'] = '128'
+
 import numpy as np
+import cupy as cp
 from numba.core.decorators import jit
 import scipy.constants as sc
 from mpi4py import MPI
@@ -27,16 +31,24 @@ from .supported_opac import supported_species, supported_cia, inactive_species
 from .parameters import assign_free_params, generate_state, \
                         unpack_geometry_params, unpack_cloud_params
 from .absorption import opacity_tables, store_Rayleigh_eta_LBL, extinction, \
-                        extinction_LBL
+                        extinction_LBL, extinction_GPU
 from .geometry import atmosphere_regions, angular_grids
 from .atmosphere import profiles
 from .instrument import init_instrument
 from .transmission import TRIDENT
-from .emission import emission_rad_transfer, determine_photosphere_radii
+from .emission import emission_rad_transfer, determine_photosphere_radii, \
+                      emission_rad_transfer_GPU, determine_photosphere_radii_GPU
 
 
 comm = MPI.COMM_WORLD
 rank = comm.Get_rank()
+
+block = int(os.environ['block'])
+thread = int(os.environ['thread'])
+
+import warnings
+
+warnings.filterwarnings("ignore") # Suppress numba performance warning
 
 
 def create_star(R_s, T_eff, log_g, Met, T_eff_error = 100.0, 
@@ -748,7 +760,8 @@ def check_atmosphere_physical(atmosphere, opac):
 def compute_spectrum(planet, star, model, atmosphere, opac, wl,
                      spectrum_type = 'transmission', save_spectrum = False,
                      disable_continuum = False, suppress_print = False,
-                     Gauss_quad = 2, use_photosphere_radius = True):
+                     Gauss_quad = 2, use_photosphere_radius = True,
+                     device = 'cpu'):
     '''
     Calculate extinction coefficients, then solve the radiative transfer 
     equation to compute the spectrum of the model atmosphere.
@@ -908,21 +921,61 @@ def compute_spectrum(planet, star, model, atmosphere, opac, wl,
         # Also unpack fine temperature and pressure grids from pre-interpolation
         T_fine = opac['T_fine']
         log_P_fine = opac['log_P_fine']
+
+        # Running POSEIDON on the CPU
+        if (device == 'cpu'):
         
-        # Calculate extinction coefficients in standard mode
-        kappa_clear, kappa_cloud = extinction(chemical_species, active_species,
-                                              CIA_pairs, ff_pairs, bf_species,
-                                              n, T, P, wl, X, X_active, X_CIA, 
-                                              X_ff, X_bf, a, gamma, P_cloud, 
-                                              kappa_cloud_0, sigma_stored, 
-                                              CIA_stored, Rayleigh_stored, 
-                                              ff_stored, bf_stored, enable_haze, 
-                                              enable_deck, enable_surface,
-                                              N_sectors, N_zones, T_fine, 
-                                              log_P_fine, P_surf)
+            # Calculate extinction coefficients in standard mode
+            kappa_clear, kappa_cloud = extinction(chemical_species, active_species,
+                                                  CIA_pairs, ff_pairs, bf_species,
+                                                  n, T, P, wl, X, X_active, X_CIA, 
+                                                  X_ff, X_bf, a, gamma, P_cloud, 
+                                                  kappa_cloud_0, sigma_stored, 
[…]
