### 2022-10-15 5631a42
Added xsec arrays defined on GPU
