### 2022-10-19 370e69d
Initial gradient parameter test
