+++ POSEIDON/geometry.py
@@ -4,6 +4,8 @@ Functions to compute geometric aspects of the model atmosphere.
 '''
 
 import numpy as np
+import scipy.constants as sc
+from scipy.special import erfinv
 from numba.core.decorators import jit
 
 
@@ -209,3 +211,71 @@ def angular_grids(Atmosphere_dimension, TwoD_type, N_slice_EM, N_slice_DN,
     
     return phi, theta, phi_edge, theta_edge, dphi, dtheta
     
+
+def opening_angle(R_p_ref, T, mu, g, log_sigma = -25.3, log_X = -3.3, 
+                  tau_0 = 0.56, P_ref = 0.01, gamma = 0.01):
+    '''
+    Calculate the predicted terminator opening angle using the analytic equation
+    from Wardenier+2022 (their Eq. 10).
+
+    This analytic prescription is valid for a 1D isothermal atmosphere with
+    constant gravity with opacity dominated by a single chemical species with
+    constant cross section. Despite these caveats, the formula performs rather
+    well compared to full radiative transfer models. Therefore, this offers
+    a good first-order estimate for the angular width about the terminator
+    plane for which transmission spectrum are sensitive.  
+
+    Args:
+        R_p_ref (float):
+            Planetary radius at the reference pressure P_ref (typically the 
+            white light transit radius).
+        T (float):
+            Representative planet temperature (typically T_eq) (K).
+        mu (float):
+            Mean molecular mass of the atmosphere (kg).
+        g (float):
+            Gravitational field strength of planet (m/s^2).
+        log_sigma (float):
+            Absorption cross section of dominant opacity source (log_10 (m^2)).
+        log_X (float):
+            Mixing ratio of chemical species dominating the opacity.
+        tau_0 (float):
+            Slant optical depth for impact parameter corresponding to the opening
+            angle (0.56 default corresponds to Lecavelier des Etangs+2008).
+        P_ref (float):
+            Reference pressure corresponding to R_p_ref.
+        gamma (float):
+            'Tolerance factor' for the opening angle, such that the angle 
+            encloses a fraction gamma -> (1 - gamma) of the total slant optical
+            depth. The default corresponds to the 1-99 % optical depth region.
+            Note: relabelled from 'beta' in Wardenier+2022, since POSEIDON uses
+            beta for the opening angle itself.
+
+    Returns:
+        beta (float):
+            Day-night opening angle (degrees).
+
+    '''
+
+    # Compute number density at reference pressure 
+    n_0 = P_ref / (sc.k * T)    # Ideal gas law
+
+    # Compute constant pressure scale height
+    H = (sc.k * T) / (mu * g)
+
+    # Convert cross section and mixing ratio to linear units
+    sigma = np.power(10.0, log_sigma)   # Cross section (m^2)
+    X = np.power(10.0, log_X)           # Mixing ratio
+
+    # Store constant encoding assumed chemical composition
+    chem_const = ((sigma * X) / tau_0)    # Default value: 10^(-28.3) m^2
+
+    # Calculate terminator opening angle
+    beta = 2.0 * np.arctan((np.sqrt(2 * R_p_ref * H) * erfinv (1.0 - gamma)) / 
+                           (R_p_ref + H * np.log(np.sqrt(2.0 * np.pi * R_p_ref * H) * chem_const * n_0)))
+
+    # Convert from radians to degrees
+    beta *= (180.0 / np.pi)
+
+    return beta
+
