+++ POSEIDON/parameters.py
@@ -204,6 +204,16 @@ def assign_free_params(param_species, object_type, PT_profile, X_profile,
                 elif (PT_profile == 'two-gradients'):            
                     PT_params += ['T_bar_DN_high', 'T_bar_DN_mid', 'Delta_T_DN_high', 
                                   'Delta_T_DN_mid', 'log_P_mid', 'T_deep']
+
+        # Gradient parameter prescription from MacDonald & Lewis (2023)
+        elif (TwoD_param_scheme == 'gradient'):
+
+            if (TwoD_type == 'D-N'):
+                if (PT_profile == 'gradient'):
+                    PT_params += ['T_bar_DN_high', 'Grad_theta_T_high', 'T_deep']
+                elif (PT_profile == 'two-gradients'):            
+                    PT_params += ['T_bar_DN_high', 'T_bar_DN_mid', 'Grad_theta_T_high', 
+                                  'Grad_theta_T_mid', 'log_P_mid', 'T_deep']
     
     # 3D model (asymmetric terminator + day-night transition)
     elif (PT_dim == 3):
@@ -333,6 +343,33 @@ def assign_free_params(param_species, object_type, PT_profile, X_profile,
                                              'log_P_' + species + '_mid', 'log_' + species + '_deep']
                         else:   # No altitude variation for this species
                             X_params += ['log_' + species + '_bar_DN', 'Delta_log_' + species + '_DN']
+
+                    else:       # No Day-Night variation for this species
+                        if ((species_vert_gradient != []) and (species in species_vert_gradient)):
+                            if (X_profile == 'gradient'):  
+                                X_params += ['log_' + species + '_high', 'log_' + species + '_deep']
+                            elif (X_profile == 'two-gradients'):  
+                                X_params += ['log_' + species + '_high', 'log_' + species + '_mid', 
+                                            'log_P_' + species + '_mid', 'log_' + species + '_deep']      
+                        else:   # No altitude variation for this species
+                            X_params += ['log_' + species]
+
+            # Gradient parameter prescription from MacDonald & Lewis (2023)
+            if (TwoD_param_scheme == 'gradient'):
+
+                # Species with variation only across the terminator (2D Day-Night X_i)
+                if (TwoD_type == 'D-N'):                
+                    if ((species_DN_gradient != []) and (species in species_DN_gradient)):
+                        if ((species_vert_gradient != []) and (species in species_vert_gradient)):
+                            if (X_profile == 'gradient'):  
+                                X_params += ['log_' + species + '_bar_DN_high', 'Grad_theta_log_' + species + '_high', 
+                                             'log_' + species + '_deep']
+                            elif (X_profile == 'two-gradients'):  
+                                X_params += ['log_' + species + '_bar_DN_high', 'log_' + species + '_bar_DN_mid',
+                                             'Grad_theta_log_' + species + '_high', 'Grad_theta_log_' + species + '_mid', 
+                                             'log_P_' + species + '_mid', 'log_' + species + '_deep']
+                        else:   # No altitude variation for this species
+                            X_params += ['log_' + species + '_bar_DN', 'Grad_theta_log_' + species + '_DN']
 
                     else:       # No Day-Night variation for this species
                         if ((species_vert_gradient != []) and (species in species_vert_gradient)):
@@ -590,7 +627,7 @@ def split_params(params_drawn, N_params_cumulative):
     
[…]

[…]
