### 2022-10-19 f50a502
Fix for cupy import errors if not installed
