+++ POSEIDON/utility.py
@@ -1501,5 +1501,10 @@ def write_MultiNest_results(planet, model, data, retrieval_name,
                        wl, R, instruments, datasets)
     
 
-
+def mock_missing(name):
+    def init(self, *args, **kwargs):
+        raise ImportError(
+            f'The module {name} you tried to call is not importable; '
+            f'this is likely due to it not being installed.')
+    return type(name, (), {'__init__': init})
 
