### 2022-10-23 82040b2
Syntax changes to match MacDonald & Lewis (2022)
