### 2022-10-23 #4873 @editorialbot
[REVIEW]: POSEIDON: A Multidimensional Atmospheric Retrieval Code for Exoplanet Spectra
**Submitting author:** @MartianColonist (<a href="http://orcid.org/0000-0003-4816-3469">Ryan J. MacDonald</a>)
**Repository:** https://github.com/MartianColonist/POSEIDON
**Branch with paper.md** (empty if default branch): main
**Version:** 1.0.0
