### 2022-10-27 #4873 @LorenzoMugnai
Hello @MartianColonist. I have a couple of comments on the paper.
The writing is clear and simple, but I feel more can be said about comparing to other software or, as it is called in the checklist, the "Field Status". You say something in the Statement of Need section, but it doesn't seem complete to me. Perhaps you can add a few words along the line of the conclusion of Barstow et al. 2022 (https://ui.adsabs.harvard.edu/abs/2022ExA....53..447B/abstract) to demonstrate that the performance of POISEIDON is comparable to others for 1D atmospheres. This will tell the reader that you not only provide 2D or 3D capabilities, but for the same exercises that your code performs as the others in the field.
I think it might be nice to mention the Nemesis (https://www.sciencedirect.com/science/article/abs/pii/S0022407307003378?via%3Dihub) and Arcis (https://www.aanda.org/articles/aa/full_html/2020/10/aa37377-19/aa37377-19.html) software as well.
