### 2022-10-27 #4873 @MartianColonist
Hi @LorenzoMugnai,

Thanks for your comments, I'll let you know once the paper has been revised.

A couple quick answers / questions:

1. There is a 1D model benchmark for POSEIDON in Appendix D of MacDonald & Lewis (2022) (https://ui.adsabs.harvard.edu/abs/2022ApJ...929...20M/abstract). The 1D model benchmark was 70 ms, which is a factor of a few faster than other transmission spectra codes I've tested against. I'll add these details to the paper.
2. Could you clarify which conclusuon of Barstow et al. 2022 you are referencing? POSEIDON was one of the codes included in that paper, so the results matched the other codes.
3. A very good point about NEMESIS and ARCiS! I was just referencing open source retrieval codes (since there are around 20 other closed source codes mentioned in the literature). But it appears NEMESIS does have a GitHub repository, so I'll be sure to reference it. ARCiS is not open source from what I can tell.
