### 2022-11-04 a388820
Better formatting for C/O function
