+++ joss/paper.bib
@@ -390,3 +390,16 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@ARTICLE{Irwin2008,
+       author = {{Irwin}, P.~G.~J. and {Teanby}, N.~A. and {de Kok}, R. and {Fletcher}, L.~N. and {Howett}, C.~J.~A. and {Tsang}, C.~C.~C. and {Wilson}, C.~F. and {Calcutt}, S.~B. and {Nixon}, C.~A. and {Parrish}, P.~D.},
+        title = "{The NEMESIS planetary atmosphere radiative transfer and retrieval tool}",
+      journal = {Journal of Quantitative Spectroscopy and Radiative Transfer},
+         year = 2008,
+        month = apr,
+       volume = {109},
+        pages = {1136-1150},
+          doi = {10.1016/j.jqsrt.2007.11.006},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2008JQSRT.109.1136I},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
