+++ joss/paper.bib
@@ -403,3 +403,76 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@ARTICLE{Lewis2020,
+       author = {{Lewis}, N.~K. and {Wakeford}, H.~R. and {MacDonald}, R.~J. and {Goyal}, J.~M. and {Sing}, D.~K. and {Barstow}, J. and {Powell}, D. and {Kataria}, T. and {Mishra}, I. and {Marley}, M.~S. and {Batalha}, N.~E. and {Moses}, J.~I. and {Gao}, P. and {Wilson}, T.~J. and {Chubb}, K.~L. and {Mikal-Evans}, T. and {Nikolov}, N. and {Pirzkal}, N. and {Spake}, J.~J. and {Stevenson}, K.~B. and {Valenti}, J. and {Zhang}, X.},
+        title = "{Into the UV: The Atmosphere of the Hot Jupiter HAT-P-41b Revealed}",
+      journal = {Astrophysical Journal Letters},
+     keywords = {Exoplanet atmospheres, Observational astronomy, Exoplanet atmospheric composition, Spectroscopy, 487, 1145, 2021, 1558, Astrophysics - Earth and Planetary Astrophysics},
+         year = 2020,
+        month = oct,
+       volume = {902},
+       number = {1},
+          eid = {L19},
+        pages = {L19},
+          doi = {10.3847/2041-8213/abb77f},
+archivePrefix = {arXiv},
+       eprint = {2010.08551},
+ primaryClass = {astro-ph.EP},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2020ApJ...902L..19L},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{Rathcke2021,
+       author = {{Rathcke}, Alexander D. and {MacDonald}, Ryan J. and {Barstow}, Joanna K. and {Goyal}, Jayesh M. and {Lopez-Morales}, Mercedes and {Mendon{\c{c}}a}, Jo{\~a}o M. and {Sanz-Forcada}, Jorge and {Henry}, Gregory W. and {Sing}, David K. and {Alam}, Munazza K. and {Lewis}, Nikole K. and {Chubb}, Katy L. and {Taylor}, Jake and {Nikolov}, Nikolay and {Buchhave}, Lars A.},
+        title = "{HST PanCET Program: A Complete Near-UV to Infrared Transmission Spectrum for the Hot Jupiter WASP-79b}",
+      journal = {Astronomical Journal},
+     keywords = {Exoplanet atmospheres, Exoplanet atmospheric composition, Observational astronomy, Hot Jupiters, Transmission spectroscopy, 487, 2021, 1145, 753, 2133, Astrophysics - Earth and Planetary Astrophysics},
+         year = 2021,
+        month = oct,
+       volume = {162},
+       number = {4},
+          eid = {138},
+        pages = {138},
+          doi = {10.3847/1538-3881/ac0e99},
+archivePrefix = {arXiv},
+       eprint = {2104.10688},
+ primaryClass = {astro-ph.EP},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2021AJ....162..138R},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{Barstow2022,
+       author = {{Barstow}, Joanna K. and {Changeat}, Quentin and {Chubb}, Katy L. and {Cubillos}, Patricio E. and {Edwards}, Billy and {MacDonald}, Ryan J. and {Min}, Michiel and {Waldmann}, Ingo P.},
+        title = "{A retrieval challenge exercise for the Ariel mission}",
[…]
