+++ joss/paper.bib
@@ -390,7 +390,7 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@ARTICLE{Irwin2008,
+@ARTICLE{Irwin:2008,
        author = {{Irwin}, P.~G.~J. and {Teanby}, N.~A. and {de Kok}, R. and {Fletcher}, L.~N. and {Howett}, C.~J.~A. and {Tsang}, C.~C.~C. and {Wilson}, C.~F. and {Calcutt}, S.~B. and {Nixon}, C.~A. and {Parrish}, P.~D.},
         title = "{The NEMESIS planetary atmosphere radiative transfer and retrieval tool}",
       journal = {Journal of Quantitative Spectroscopy and Radiative Transfer},
@@ -403,7 +403,7 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@ARTICLE{Lewis2020,
+@ARTICLE{Lewis:2020,
        author = {{Lewis}, N.~K. and {Wakeford}, H.~R. and {MacDonald}, R.~J. and {Goyal}, J.~M. and {Sing}, D.~K. and {Barstow}, J. and {Powell}, D. and {Kataria}, T. and {Mishra}, I. and {Marley}, M.~S. and {Batalha}, N.~E. and {Moses}, J.~I. and {Gao}, P. and {Wilson}, T.~J. and {Chubb}, K.~L. and {Mikal-Evans}, T. and {Nikolov}, N. and {Pirzkal}, N. and {Spake}, J.~J. and {Stevenson}, K.~B. and {Valenti}, J. and {Zhang}, X.},
         title = "{Into the UV: The Atmosphere of the Hot Jupiter HAT-P-41b Revealed}",
       journal = {Astrophysical Journal Letters},
@@ -422,7 +422,7 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@ARTICLE{Rathcke2021,
+@ARTICLE{Rathcke:2021,
        author = {{Rathcke}, Alexander D. and {MacDonald}, Ryan J. and {Barstow}, Joanna K. and {Goyal}, Jayesh M. and {Lopez-Morales}, Mercedes and {Mendon{\c{c}}a}, Jo{\~a}o M. and {Sanz-Forcada}, Jorge and {Henry}, Gregory W. and {Sing}, David K. and {Alam}, Munazza K. and {Lewis}, Nikole K. and {Chubb}, Katy L. and {Taylor}, Jake and {Nikolov}, Nikolay and {Buchhave}, Lars A.},
         title = "{HST PanCET Program: A Complete Near-UV to Infrared Transmission Spectrum for the Hot Jupiter WASP-79b}",
       journal = {Astronomical Journal},
@@ -441,7 +441,7 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@ARTICLE{Barstow2022,
+@ARTICLE{Barstow:2022,
        author = {{Barstow}, Joanna K. and {Changeat}, Quentin and {Chubb}, Katy L. and {Cubillos}, Patricio E. and {Edwards}, Billy and {MacDonald}, Ryan J. and {Min}, Michiel and {Waldmann}, Ingo P.},
         title = "{A retrieval challenge exercise for the Ariel mission}",
       journal = {Experimental Astronomy},
@@ -459,7 +459,7 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
[…]
