+++ joss/paper.md
@@ -47,7 +47,7 @@ Recent years have seen a substantial improvement in the number of high-quality e
 
 `POSEIDON` provides the capability to model and retrieve planets with inhomogeneous temperatures, compositions, and cloud properties (i.e. 2D or 3D models). Several studies have highlighted that not including these multidimensional effects can bias retrieval inferences [e.g., @Line:2016;  @Caldas:2019; @MacDonald:2020; @Pluriel:2022]. However, existing open-source exoplanet retrieval codes assume 1D atmospheres for computational efficiency. `POSEIDON` therefore offers an open-source implementation of state-of-the-art multidimensional retrieval methods [see @MacDonald:2022 and MacDonald & Lewis (in prep.)] to aid the interpretation of high-quality exoplanet spectra.
 
-In a 1D configuration, `POSEIDON` compares well with other retrieval codes. When applied to Hubble Space Telescope observations, `POSEIDON` produces consistent retrieval results with the ATMO and NEMESIS retrieval codes [@Lewis:2020; @Rathcke:2021]. Recently, [@Barstow:2022] presented a comparison of five exoplanet retrieval codes, including `POSEIDON`, which demonstrated good agreement on simulated Ariel [@Tinetti2020] transmission spectra. `POSEIDON` also offers exceptional computational performance: a single 1D forward model over a wavelength range sufficient for JWST analyses takes 70 ms [see @MacDonald:2022, Appendix D], while publication-quality 1D retrievals typically take an hour or less. `POSEIDON` also supports multi-core retrievals via `PyMultiNest`'s MPI implementation, which achieves a roughly linear speed-up in the number of cores. Therefore, `POSEIDON` allows users to readily explore 1D retrievals on personal laptops while scaling up to multidimensional retrievals on modest clusters.
+In a 1D configuration, `POSEIDON` compares well with other retrieval codes. When applied to Hubble Space Telescope observations, `POSEIDON` produces consistent retrieval results with the ATMO and NEMESIS retrieval codes [@Lewis:2020; @Rathcke:2021]. Recently, [@Barstow:2022] presented a comparison of five exoplanet retrieval codes, including `POSEIDON`, which demonstrated good agreement on simulated Ariel [@Tinetti:2020] transmission spectra. `POSEIDON` also offers exceptional computational performance: a single 1D forward model over a wavelength range sufficient for JWST analyses takes 70 ms [see @MacDonald:2022, Appendix D], while publication-quality 1D retrievals typically take an hour or less. `POSEIDON` also supports multi-core retrievals via `PyMultiNest`'s MPI implementation, which achieves a roughly linear speed-up in the number of cores. Therefore, `POSEIDON` allows users to readily explore 1D retrievals on personal laptops while scaling up to multidimensional retrievals on modest clusters.
 
 # Documentation
 
