### 2022-11-04 #4873 @MartianColonist
Hi @LorenzoMugnai,

I've updated the paper to address your feedback. There is now a new paragraph at the end of the 'statement of need' section comparing POSEIDON to other codes for the 1D case.

Thanks for the suggestion!
