### 2022-11-07 #4873 @MartianColonist
Hi @LorenzoMugnai,

Regarding "Community guidelines", all there is at present is the brief 'support' section at the end of the GitHub README. I will add a dedicated section on community contribution guidelines to the documentation to improve this. I'll let you know once it's up.

For tests, it isn't obvious how to concoct future-proof automated tests for a radiative transfer and retrieval package like POSEIDON. For automated tests, two issues come to mind:

1. If a 'reference spectrum' is used in tests, future updates to the cross sections / opacity files would cause the test to fail (due to small changes in the line intensities / positions). Since new molecular line lists are frequently released, this would raise a test failure even if none of the radiative transfer code has changed.
2. A posterior, being statistical in nature, will be slightly different every time POSEIDON is run. So there isn't an obvious way to have a reference retrieval result in an automated test.

So I opted for a manual test approach: the tutorials in the documentation serve to verify POSEIDON's functionality. A user can download a Jupyter notebook for each tutorial in the guide by clicking the 'Download full notebook here' link at the top of each page. Since the expected output is included in the plots built into each Jupyter notebook, by successfully running the notebook cells the user can verify the functionality.

Per the JOSS review criteria, this is essentially:

_OK: Documented manual steps that can be followed to objectively check the expected functionality of the software (e.g., a sample input file to assert behavior)_
