### 2022-11-07 #4873 @ideasrule
Unit testing really is tricky for retrieval codes!  It's not mandated by JOSS, but I do recommend having some rudimentary unit tests, if only to catch obvious bugs and set up the framework for adding more tests later on.  It'll give you some peace of mind when you make major revisions to the code.

Here are some suggestions for unit tests:

1. Your code supports 1D, 2D, and 3D atmospheres.  Computing the transit spectrum for a 2D atmosphere with the inhomogeneity set to 0 should give you the same result as computing a 1D atmosphere, and similarly a 3D atmosphere with azimuthal symmetry should be the same as a 2D atmosphere.  Is this the case?

2. There are some simple cases where you can analytically calculate the transit spectrum.  For example, for an isothermal atmosphere with H << Rp and a cross section that doesn't depend on height, the transit depth is given by Equation 27 in Betremieux & Swain 2017: https://academic.oup.com/mnras/article/467/3/2834/2982874

I used it to compute the transit depth if a Rayleigh scattering haze were the only source of opacity (Eq 11-13 in this paper: https://arxiv.org/pdf/1811.11761.pdf)

Since the formula is linear in H and ln(abundance*sigma), it ought to be pretty easy to generalize to a temperature and/or ln(abundance) distribution that varies linearly with latitude.

3. You probably calculate the slant optical depth somewhere in the code.  Does the slant optical depth tau_slant(P) equal tau_vert(P) * sqrt(2*pi*Rs/H) = H*n(P)*sigma*sqrt(2*pi*Rs/H), where n is the number density at pressure P, Rs is the planet radius at that pressure, sigma is the cross section, and H is the scale height?  (This is all assuming an isothermal atmosphere where sigma is constant and H << Rs.)  Also, if you take the slant optical depth that you calculate and apply the analytic formula (Betremieux & Swain 2017), do you get the transit depth that your code calculates?

4. It's very hard to test if a retrieval is giving you the right results, but you could put it in a unit test and force it to terminate after some small number of iterations, just to make sure the retrieval actually runs and doesn't crash immediately.
