### 2022-11-07 #4873 @MartianColonist
These are good suggestions, thanks @ideasrule!

I'll look into some unit tests along the lines above.
