### 2022-11-12 #4873 @ideasrule
So far, I've successfully installed POSEIDON (a very painless process) and ran transmission_basic.ipynb and transmission_2D.ipynb without problem.  However, transmission_terrestrial.ipynb runs into errors, possibly related to the open issue you have on Github:  [
![Screenshot from 2022-11-11 19-03-55](https://user-images.githubusercontent.com/1724065/201449172-9870f576-b713-4681-8058-efbad4746743.png)
](url)

Could you see what might be wrong?

I also had a quick comment on the paper.  Looking at the line lists used to generate opacities, they mostly seem to be up to date and of high quality, but I notice that CH4 uses ExoMol's 34to10.  This is good enough for low resolution work, but the line positions in this line list are not accurate enough for high resolution cross correlation spectroscopy.  Would you consider switching to HITEMP 2020's line list?  From their paper (https://iopscience.iop.org/article/10.3847/1538-4365/ab7a1a):

"The completeness of the 34to10 line list has improved when compared to 10to10, with the partitioning of the line list making it more practical to use. However, the underlying energy levels (and transition frequencies) have not been adjusted, and therefore, the accuracy issues noted for 10to10 remain relevant to 34to10. Line intensities are also significantly overestimated with respect to experimental data for high wavenumber ranges."
