### 2022-11-13 eafd254
Removed superfluous argument
