### 2022-11-13 273f5aa
Removed superfluous argument
