### 2022-11-13 #4873 @MartianColonist
Hi @ideasrule,

The typo in 'transmission_terrestrial.ipynb' has been fixed on the main branch. Can you verify it works for you now?

Regarding the line lists, I'm part of a collaboration working on a NASA-supported opacity database (https://ui.adsabs.harvard.edu/abs/2022BAAS...54e.371B/abstract). Once this database is publicly released, I will comprehensively update all POSEIDON's cross sections (including CH4) to the latest line lists recommended by the ExoMol and HITRAN/HITEMP teams (who are both involved in this project). For the time being, the 34to10 CH4 cross section is more than sufficient for low spectral resolution JWST analyses (I don't find any significant differences in forward model comparisons vs. other codes). But everything will be updated when our new database is released.

I also note that since I am travelling for the next 3 weeks, I might be a little slow updating the code base in response to your feedback. Rest assured that I'm noting all this down, and I greatly appreciate all your feedback!
