+++ POSEIDON/high_res.py
@@ -1,0 +1,560 @@
+# Functions related to atmospheric retrieval
+
+import numpy as np
+import time
+import os
+import pymultinest
+from mpi4py import MPI
+from scipy.special import ndtri
+from spectres import spectres
+from numba.core.decorators import jit
+from scipy.constants import parsec
+
+from .constants import R_J, R_E
+
+from .parameters import split_params
+from .instrument import bin_spectrum_to_data
+from .utility import write_MultiNest_results, round_sig_figs, closest_index, \
+                     write_retrieved_spectrum, write_retrieved_PT, \
+                     write_retrieved_log_X, confidence_intervals
+from .core import make_atmosphere, compute_spectrum
+from .stellar import precompute_stellar_spectra, stellar_contamination_single_spot
+from .cross_correlate import log_likelihood
+
+comm = MPI.COMM_WORLD
+rank = comm.Get_rank()
+
+# Create global variable needed for centred log-ratio prior function
+allowed_simplex = 1
+
+
+
+def run_high_res_retrieval(planet, star, model, opac, data, priors, 
+                  wl, P, P_ref = 10.0, R = None, retrieval_name = None,
+                  He_fraction = 0.17, N_slice_EM = 2, N_slice_DN = 4, 
+                  spectrum_type = 'direct_emission', N_live = 400, ev_tol = 0.5,
+                  sampling_algorithm = 'MultiNest', resume = False, 
+                  verbose = True, sampling_target = 'parameter',
+                  N_output_samples = 1000):
+    '''
+    ADD DOCSTRING
+    '''
+
+    # Unpack planet name
+    planet_name = planet['planet_name']
+    
+    # Unpack prior types and ranges
+    prior_types = priors['prior_types']
+    prior_ranges = priors['prior_ranges']
+
+    # Unpack model properties
+    model_name = model['model_name']
+    chemical_species = model['chemical_species']
+    param_names = model['param_names']
+    N_params = len(param_names)
+
+    if (retrieval_name is None):
+        retrieval_name = model_name
+    else:
+        retrieval_name = model_name + '_' + retrieval_name
+
+    # Identify output directory location
+    output_dir = './POSEIDON_output/' + planet_name + '/retrievals/'
+
+    if (rank == 0):
+        print("POSEIDON now running '" + retrieval_name + "'")
+
+    # Run POSEIDON retrieval using PyMultiNest
+    if (sampling_algorithm == 'MultiNest'):
+
+        # Change directory into MultiNest result file folder
+        os.chdir(output_dir + 'MultiNest_raw/')
+
+        # Set basename for MultiNest output files
+        basename = retrieval_name + '-'
+
+        # Begin retrieval timer
+        if (rank == 0):
+            t0 = time.perf_counter()
+
+        # Run MultiNest
+        high_res_retrieval(planet, star, model, opac, data, prior_types, 
+                              prior_ranges, spectrum_type, wl, P, P_ref, 
+                              He_fraction, N_slice_EM, N_slice_DN, N_params, 
+                              resume = resume, verbose = verbose,
+                              outputfiles_basename = basename, 
+                              n_live_points = N_live, multimodal = False,
+                              evidence_tolerance = ev_tol, log_zero = -1e90,
+                              importance_nested_sampling = False, 
+                              sampling_efficiency = sampling_target, 
+                              const_efficiency_mode = False)
+
+        # Write retrieval results to file
+        if (rank == 0):
+            
+            # pdb.set_trace()
+
+            # Write retrieval runtime to terminal
+            t1 = time.perf_counter()
+            total = round_sig_figs((t1-t0)/3600.0, 2)  # Round to 2 significant figures
+            
+            print('POSEIDON retrieval finished in ' + str(total) + ' hours')
+      
[…]
