+++ POSEIDON/cross_correlate.py
@@ -1,0 +1,162 @@
+from __future__ import absolute_import, unicode_literals, print_function
+import numpy as np
+import pickle
+from scipy import constants
+from numba import jit
+from astropy.io import fits
+from scipy import interpolate
+from .core import create_star, create_planet, define_model, make_atmosphere, read_opacities, wl_grid_constant_R, wl_grid_line_by_line, compute_spectrum
+from .constants import R_Sun
+from .visuals import plot_stellar_flux
+from .constants import R_J, M_J
+import numpy as np
+from spectres import spectres
+from sklearn.decomposition import TruncatedSVD
+
+@jit
+def get_rot_kernel(V_sin_i, wl_s):
+    dRV = np.mean(2.0*(wl_s[1: ]-wl_s[0: -1])/(wl_s[1: ]+wl_s[0: -1]))*2.998E5
+    n_ker = 401
+    half_n_ker = (n_ker - 1)//2
+    rot_ker = np.zeros(n_ker)
+    for ii in range(n_ker):
+        ik = ii - half_n_ker
+        x = ik * dRV / V_sin_i
+        if np.abs(x) < 1.0:
+            y = np.sqrt(1 - x**2)
+            rot_ker[ii] = y
+    rot_ker /= rot_ker.sum()
+
+    return rot_ker
+
+
+def log_likelihood_PCA(V_sys, K_p, scale, cs_p, cs_s, wl_grid, data_arr, data_scale, V_bary, Phi):
+
+    K_s = 0.3229
+
+    #Kstar=(Mp/Mstar*9.55E-4)*Kp  #this is mass planet/mass star
+    Ndet, Nphi, Npix = data_arr.shape
+
+    I = np.ones(Npix)
+    N = Npix # np.array([Npix])
+    
+    # Time-resolved total radial velocity
+    RV_p = V_sys + V_bary + K_p * np.sin(2 * np.pi * Phi)  # Vsys is an additive term around zero   
+    dl_p = RV_p * 1e3 / constants.c
+    RV_s = (V_sys + V_bary - K_s * np.sin(2 * np.pi * Phi))*0  # Velocity of the star is very small compared to planet's velocity and it's already be corrected
+    dl_s = RV_s * 1e3 / constants.c
+    # Initializing log-likelihoods and CCFs
+    logL_Matteo = 0
+    logL_Zuck = 0
+    CCF = 0
+    # Looping through each phase and computing total log-L by summing logLs for each obvservation/phase
+    for j in range(Ndet): # Ndet = 44 This takes 2.2 seconds to complete
+        wl_slice = wl_grid[j, ].copy() # Cropped wavelengths    
+        Fp_Fs = np.zeros((Nphi, Npix))  # "shifted" model spectra array at each phase
+        for i in range(Nphi): # This for loop takes 0.025 seconds Nphi = 79
+            wl_shifted_p = wl_slice * (1.0 - dl_p[i])
+            Fp = interpolate.splev(wl_shifted_p, cs_p, der=0) * scale
+            wl_shifted_s = wl_slice * (1.0 - dl_s[i])
+            Fs = interpolate.splev(wl_shifted_s, cs_s, der=0)
+            Fp_Fs[i, :] = Fp/Fs
+
+        model_injected = (1 + Fp_Fs) * data_scale[j, :]  # 1??+fp/fstar is same as (fstar+fp)/fstar..tell "stretches" by transmittance
+        
+        # I think of this as:
+        # data_scale contains the first four principle component, data_arr contains the rest. data_arr is telluric subtracted, 
[…]

[…]
