### 2022-11-16 #4873 @ideasrule
@MartianColonist I'm actually glad to hear that you're travelling, because I was worried that I was taking too long with the review :)  Thanks for correcting the typo in transmission_terestrial.ipynb.  I can verify that it works now, and so do the other notebooks.

To further test the code, I decided to set up a retrieval on MIRI/LRS data.  I think the retrieval worked well, but the plotting ran into a minor issue: I tried to set wl_min to 4.5 and wl_max to 12.5, but the plot's x-axis is always 1, which squishes all of the useful range into the rightmost 40% of the plot.  I got around this problem by adding "plt.xlim(wl_min, wl_max) right in front of plt.tight_layout() in visuals.py:plot_spectra_retrieved, but there's surely a better way to fix it.

Another minor issue: in retrieval_basic.ipynb, data_dir is 'Tutorial/WASP-999b', but the directory doesn't actually exist.  I spent a while looking for it in order to see what the format of the data file is supposed to be, before looking at the source code and realizing it searches for that string and replaces it with a different path.  Is it possible to give it the path directly?  I suspect lots of people will try to use "Tutorial/WASP-999b" as an example data directory, just as I did, and get confused about why it doesn't exist.
