### 2022-11-16 #4873 @ideasrule
Aside from the minor issues I mentioned in the last post, I have one major comment and one more minor issue.

**Major comment**: Could you indicate clearly in the article proof, the GitHub README, and the readthedocs home page what the state of support for emission spectroscopy is?  The article proof only mentions transmission spectroscopy, the README doesn't mention either transmission or emission, and the readthedocs mentions only transmission in the first paragraph but has a section in the "Guide" on emission.  Are emission forward modelling & retrieval official features of the package (which would mean it falls under this review), or are they beta features?

**Minor issue**: It would be good to have API documentation for the parameters in the [visuals class](https://poseidon-retrievals.readthedocs.io/en/latest/autoapi/POSEIDON/visuals/index.html) methods, since almost every user will be using these methods routinely.  I'm thinking particularly of plot_spectra_retrieved and plot_spectra, and particularly of parameters where the meaning isn't obvious (e.g. what are the options for y_unit)?
