### 2022-11-16 #4873 @MartianColonist
@ideasrule emission spectra are a beta feature added post JOSS submission while I was waiting for an editor to be assigned. It is considered a beta feature because there is not currently a methods paper describing the forward model in the literature. I will add clear statements to the next paper and documentation update.

Yes, docstrings for visuals.py are on the to-do list. Thanks!
