### 2022-11-16 66147ac
Added Docstring in cross_correlate.py
