+++ POSEIDON/cross_correlate.py
@@ -31,7 +31,47 @@ def get_rot_kernel(V_sin_i, wl_s):
 
 
 def log_likelihood_PCA(V_sys, K_p, scale, cs_p, cs_s, wl_grid, data_arr, data_scale, V_bary, Phi):
+    '''
+    Perform the loglikelihood calculation using singular value decompositions.
+    Ndet: number of spectral order.
+    Nphi: number of time-resolved phases.
+    Npix: number of wavelengths per spectral order.
+    Typical values for (Ndet, Nphi, Npix) is (44, 79, 1848).
 
+    Args:
+        V_sys (float):
+            The system velocity (km/s) at which we do the loglikelihood calculation.
+        K_p (float):
+            The Keplerian velocity (km/s) at which we do the loglikelihood calculation.
+        Scale (float):
+            TODO: not really know what this parameter means.    
+        cs_p (np.array of float):
+            Spline representation of the observed flux of planet.
+        cs_s (np.array of float): 
+            Spline representation of the observed flux of star.
+        wl_grid (2D np.array of float):
+            2D wavelength grid of the data (Ndet x Npix). Typical size ~(44, 1848).
+        data_arr (3D np.array of float):
+            3D Array representing the top principal components removed data.
+            Shape: (Ndet x Nphi x Npix)
+        data_scale (3D np.array of float):
+            3D Array representing the top principal components of data.
+            Shape: (Ndet x Nphi x Npix)
+        V_bary (np.array of float):
+            Array of time-resolved Earth-star velocity. We have absorbed V_sys into V_bary, so V_sys = V_sys_literature + d_V_sys.
+            Shape: (Nphi, )
+        Phi (np.array of float):
+            Array of time-resolved phases.
+            Shpae (Nphi, )
+    
+    Returns:
+        logL_Matteo (float):
+            Loglikelihood given by Log(L) = -N/2 Log(s_f^2 - 2R(s) + s_g^2). Equation 9 in Brogi & Line 2019 March.
+        logL_Zack (float):
+            Loglikelihood given by Log(L) = -N/2 Log(1.0 - CC^2)). Equation 2 in Brogi & Line 2019 March.
+        CCF (float):
+            cross correlation value.
+    '''
     K_s = 0.3229
 
     #Kstar=(Mp/Mstar*9.55E-4)*Kp  #this is mass planet/mass star
@@ -49,7 +89,7 @@ def log_likelihood_PCA(V_sys, K_p, scale, cs_p, cs_s, wl_grid, data_arr, data_sc
     logL_Matteo = 0
     logL_Zuck = 0
     CCF = 0
-    # Looping through each phase and computing total log-L by summing logLs for each obvservation/phase
+    # Looping through each order and computing total log-L by summing logLs for each obvservation/order
     for j in range(Ndet): # Ndet = 44 This takes 2.2 seconds to complete
         wl_slice = wl_grid[j, ].copy() # Cropped wavelengths    
         Fp_Fs = np.zeros((Nphi, Npix))  # "shifted" model spectra array at each phase
@@ -91,7 +131,48 @@ def log_likelihood_PCA(V_sys, K_p, scale, cs_p, cs_s, wl_grid, data_arr, data_sc
 
 
 def cross_correlate(F_s_obs, F_p_obs, wl, K_p_arr, V_sys_arr, wl_grid, data_arr, data_scale, V_bary, Phi):
+    '''
+    Cross correlate at an array of Keplerian velocities and an array of centered system velocities given the observed flux.
+    Use this function to create the cross correlation plot of detection level.
[…]
