### 2022-11-16 5bf330f
Added Docstring in cross_correlate.py
