### 2022-12-20 51b6b8e
Added docstrings to atmosphere module
