+++ POSEIDON/atmosphere.py
@@ -12,21 +12,34 @@ from .utility import prior_index
 
 
 @jit(nopython = True)
-def compute_T_Madhu(P, a1, a2, log_P1, log_P2, log_P3, T_deep, P_set = 1.0e-2):
-    
-    ''' Computes the temperature profile for an atmosphere using a re-arranged
-        form of the P-T profile in Madhusudhan & Seager (2009).
-       
-        Inputs:
-        
-        P => pressure of each layer (bar)
-        PT_state => P-T profile parameters defined in Madhu & Seager (2009)
-                    in each atmospheric region
+def compute_T_Madhu(P, a1, a2, log_P1, log_P2, log_P3, T_set, P_set):
+    '''
+    Computes the temperature profile for an atmosphere using a re-arranged
+    form of the P-T profile parametrisation in Madhusudhan & Seager (2009).
 
-        Outputs:
-           
-        T => temperature of each layer (K)
-       
+    Args:
+        P (np.array of float):
+            Atmosphere pressure array (bar).
+        a1 (float):
+            Alpha_1 parameter (encodes slope in layer 1).
+        a2 (float):
+            Alpha_2 parameter encodes slope in layer 2).
+        log_P1 (float):
+            Pressure of layer 1-2 boundary.
+        log_P2 (float):
+            Pressure of inversion.
+        log_P3 (float):
+            Pressure of layer 2-3 boundary.
+        T_set (float):
+            Atmosphere temperature reference value at P = P_set (K).
+        P_set (float):
+            Pressure whether the temperature parameter T_set is defined (bar).
+    
+    Returns:
+        T (3D np.array of float):
+            Temperature of each layer as a function of pressure (K). 
+            Only the first axis is used for this 1D profile.
+    
     '''
 
     # Store number of layers for convenience
@@ -47,7 +60,7 @@ def compute_T_Madhu(P, a1, a2, log_P1, log_P2, log_P3, T_deep, P_set = 1.0e-2):
     # By default (P_set = 10 bar), so T(P_set) should be in layer 3
     if (log_P_set_i >= log_P3):
         
-        T3 = T_deep  # T_deep is the isothermal deep temperature T3 here
+        T3 = T_set  # T_deep is the isothermal deep temperature T3 here
         
         # Use the temperature parameter to compute boundary temperatures
         T2 = T3 - ((1.0/a2)*(log_P3 - log_P2))**2    
@@ -58,7 +71,7 @@ def compute_T_Madhu(P, a1, a2, log_P1, log_P2, log_P3, T_deep, P_set = 1.0e-2):
     elif (log_P_set_i >= log_P1):   # Temperature parameter in layer 2
         
         # Use the temperature parameter to compute the boundary temperatures
-        T2 = T_deep - ((1.0/a2)*(log_P_set_i - log_P2))**2  
+        T2 = T_set - ((1.0/a2)*(log_P_set_i - log_P2))**2  
         T1 = T2 + ((1.0/a2)*(log_P1 - log_P2))**2   
         T3 = T2 + ((1.0/a2)*(log_P3 - log_P2))**2
         T0 = T1 - ((1.0/a1)*(log_P1 - log_P_min))**2   
@@ -67,7 +80,7 @@ def compute_T_Madhu(P, a1, a2, log_P1, log_P2, log_P3, T_deep, P_set = 1.0e-2):
     elif (log_P_set_i < log_P1):  # Temperature parameter in layer 1
     
[…]
