### 2022-12-20 f0a3fbb
Added general element ratio function
