+++ POSEIDON/atmosphere.py
@@ -969,219 +969,159 @@ def compute_mean_mol_mass(P, X, N_species, N_sectors, N_zones, masses_all):
 
 #***** TBD: replace functions below with a general function for any elemental ratio *****#
 
-
-def locate_X(X, species, included_species):
-    
-    ''' Finds the mixing ratio for a specified species. Returns zero if
-        specified species is not included in this model.
-    
-        Inputs:
-            
-        X => volume mixing ratios of atmosphere
-        species => string giving desired chemical species
-        included_species => array of strings listing chemical species included in model
-        
-        Outputs:
-            
-        X_species => mixing ratio for this species
-    
-    '''   
-    
-    if (species in included_species):
-        X_species = X[included_species == species]
-    else:
-        X_species = 0.0
-        
-    return X_species
-
-
-def compute_metallicity(X, all_species):
-    
-    ''' Computes the metallicity [ (O/H)/(O/H)_solar ] of the atmosphere.
-    
-        Inputs:
-            
-        X => volume mixing ratios of atmosphere
-        all_species => array of strings listing chemical species included in model
-        
-        Outputs:
-            
-        M => metallicity of atmosphere
-    
+def count_atoms(molecule):
     '''
-    
-    O_to_H_solar = np.power(10.0, (8.69-12.0))  # Asplund (2009) ~ 4.9e-4  (Present day photosphere value)
-    
-    X_H2 = locate_X(X, 'H2', all_species)
-    X_H2O = locate_X(X, 'H2O', all_species)
-    X_CH4 = locate_X(X, 'CH4', all_species)
-    X_NH3 = locate_X(X, 'NH3', all_species)
-    X_HCN = locate_X(X, 'HCN', all_species)
-    X_CO = locate_X(X, 'CO', all_species)
-    X_CO2 = locate_X(X, 'CO2', all_species)
-    X_C2H2 = locate_X(X, 'C2H2', all_species)
-    X_TiO = locate_X(X, 'TiO', all_species)
-    X_VO = locate_X(X, 'VO', all_species)
-    X_AlO = locate_X(X, 'AlO', all_species)
-    X_CaO = locate_X(X, 'CaO', all_species)
-    X_TiH = locate_X(X, 'TiH', all_species)
-    X_CrH = locate_X(X, 'CrH', all_species)
-    X_FeH = locate_X(X, 'FeH', all_species)
-    X_ScH = locate_X(X, 'ScH', all_species)
-    
-    O_to_H = ((X_H2O + X_CO + 2*X_CO2 + X_TiO + X_VO + X_AlO + X_CaO)/
-              (2*X_H2 + 2*X_H2O + 4*X_CH4 + 3*X_NH3 + X_HCN + +2*X_C2H2 + X_TiH + X_CrH + X_FeH + X_ScH))
-        
-    M = O_to_H / O_to_H_solar
-    
-    return M
+    Count how many atoms of each element are contained in a molecule.
 
+    Args:
+        molecule (str):
+            Name of the molecule (e.g. 'H2O', 'CaOH', 'HC(CH3)3').
 
-def compute_C_to_O(X, all_species):
-    
-    ''' Computes the carbon-to-oxygen ratio of the atmosphere.
-    
-        Inputs:
-            
-        X => volume mixing ratios of atmosphere
-        all_species => array of strings listing chemical species included in model
-        
-        Outputs:
-            
-        C_to_O => C/O ratio of atmosphere
-    
+    Returns:
+        counts (dict):
+            Dictionary containing element counts (e.g. for H2O: {'H': 2, 'O': 1}).
     '''
-    
[…]
