### 2022-12-20 bd8be30
Finished cleaning atmosphere module docs
