### 2022-12-20 0b5ce06
Standardised cornerplot docs
