+++ POSEIDON/corner.py
@@ -12,11 +12,6 @@ Ryan MacDonald [POSEIDON modifications] (2021-2022)
 from __future__ import absolute_import, unicode_literals, print_function, division
 
 import numpy
-
-# code from dynesty (MIT licensed)
-
-#from six.moves import range
-
 import logging
 import types
 import math
@@ -46,81 +41,78 @@ SQRTEPS = math.sqrt(float(np.finfo(np.float64).eps))
 comm = MPI.COMM_WORLD
 rank = comm.Get_rank()
 
-def _quantile(x, q, weights=None):
-    """
-    Compute (weighted) quantiles from an input set of samples.
-    Parameters
-    ----------
-    x : `~numpy.ndarray` with shape (nsamps,)
-        Input samples.
-    q : `~numpy.ndarray` with shape (nquantiles,)
-       The list of quantiles to compute from `[0., 1.]`.
-    weights : `~numpy.ndarray` with shape (nsamps,), optional
-        The associated weight from each sample.
-    Returns
-    -------
-    quantiles : `~numpy.ndarray` with shape (nquantiles,)
-        The weighted sample quantiles computed at `q`.
-    """
 
-    # Initial check.
+def _quantile(x, q, weights=None):
+    '''
+    Compute (weighted) quantiles from an input set of samples.
+
+    Args:
+        x (np.array of float):
+            Input samples.
+        q (np.array of float): 
+            The list of quantiles to compute (ranging from 0 to 1).
+        weights (np.array of float):
+            The associated weight from each sample.
+    
+    Returns:
+        quantiles (np.array of float):
+            The weighted sample quantiles computed at 'q'.
+    
+    '''
+
+    # Initial check
     x = np.atleast_1d(x)
     q = np.atleast_1d(q)
 
-    # Quantile check.
+    # Quantile check
     if np.any(q < 0.0) or np.any(q > 1.0):
         raise ValueError("Quantiles must be between 0. and 1.")
 
+    # If no weights provided, this simply calls 'np.percentile'
     if weights is None:
-        # If no weights provided, this simply calls `np.percentile`.
+
         return np.percentile(x, list(100.0 * q))
+
+    # If weights are provided, compute the weighted quantiles
     else:
-        # If weights are provided, compute the weighted quantiles.
         weights = np.atleast_1d(weights)
         if len(x) != len(weights):
             raise ValueError("Dimension mismatch: len(weights) != len(x).")
-        idx = np.argsort(x)  # sort samples
-        sw = weights[idx]  # sort weights
-        cdf = np.cumsum(sw)[:-1]  # compute CDF
-        cdf /= cdf[-1]  # normalize CDF
-        cdf = np.append(0, cdf)  # ensure proper span
+        idx = np.argsort(x)              # Sort samples
+        sw = weights[idx]                # Sort weights
+        cdf = np.cumsum(sw)[:-1]         # Compute CDF
+        cdf /= cdf[-1]                   # Normalise CDF
+        cdf = np.append(0, cdf)          # Ensure proper span
         quantiles = np.interp(q, cdf, x[idx]).tolist()
+
         return quantiles
 
 
 def resample_equal(samples, weights, rstate=None):
-    """
-    Resample a new set of points from the weighted set of inputs
-    such that they all have equal weight.
+    '''
+    Resample a new set of points from the weighted set of inputs, such that 
+    they all have equal weight.
+
     Each input sample appears in the output array either
-    `floor(weights[i] * nsamples)` or `ceil(weights[i] * nsamples)` times,
-    with `floor` or `ceil` randomly selected (weighted by proximity).
-    Parameters
-    ----------
-    samples : `~numpy.ndarray` with shape (nsamples,)
-        Set of unequally weighted samples.
-    weights : `~numpy.ndarray` with shape (nsamples,)
[…]
