### 2022-12-20 38bfed0
Added chi^2 and dof to retrieval output files
