### 2022-12-22 6d94c7e
Improved wl tick generator + initial visuals docs
