+++ POSEIDON/visuals.py
@@ -23,20 +23,20 @@ from matplotlib import gridspec
 from mpl_toolkits.axes_grid1.inset_locator import inset_axes
 
 plt.style.use('classic')
-plt.rc('font', family='serif')
+plt.rc('font', family = 'serif')
 matplotlib.rcParams['svg.fonttype'] = 'none'
-matplotlib.rcParams['figure.facecolor']='white'
+matplotlib.rcParams['figure.facecolor'] = 'white'
 
 import warnings
 
 # Suppress invalid warning about x limits which sometimes occur for spectra plots
-warnings.filterwarnings("ignore", message="Attempted to set non-positive left " + 
-                                          "xlim on a log-scaled axis.\n" + 
-                                          "Invalid limit will be ignored.")
+warnings.filterwarnings("ignore", message = "Attempted to set non-positive left " + 
+                                            "xlim on a log-scaled axis.\n" + 
+                                            "Invalid limit will be ignored.")
 
-warnings.filterwarnings("ignore", message="This figure includes Axes that are " +
-                                           "not compatible with tight_layout, " +
-                                           "so results might be incorrect.")
+warnings.filterwarnings("ignore", message = "This figure includes Axes that are " +
+                                            "not compatible with tight_layout, " +
+                                            "so results might be incorrect.")
 
 from .utility import bin_spectrum, generate_latex_param_names, round_sig_figs, \
                      confidence_intervals
@@ -60,6 +60,20 @@ linestyles = {
               }
 
 def scale_lightness(colour_name, scale):
+    ''' 
+    Scale the lightness of a colour by the given factor.
+    
+    Args:
+        colour_name (str): 
+            The name of the colour to be scaled, in matplotlib colour format.
+        scale (float): 
+            The factor by which to scale the lightness of the colour (< 1 makes
+            the colour darker).
+    
+    Returns:
+        tuple: 
+            A tuple containing the RGB values of the scaled colour.
+    '''
     
     # Convert colour name to RBG value
     rgb = matplotlib.colors.ColorConverter.to_rgb(colour_name)
@@ -68,7 +82,6 @@ def scale_lightness(colour_name, scale):
     h, l, s = colorsys.rgb_to_hls(*rgb)
     
     # Manipulate h, l, s values and return as RGB
-    
     return colorsys.hls_to_rgb(h, min(1, l * scale), s = s)
 
 
@@ -877,10 +890,25 @@ def plot_chem(planet, model, atmosphere, plot_species = [],
 def set_spectrum_wl_ticks(wl_min, wl_max, wl_axis = 'log'):
     '''
     Calculates default x axis tick spacing for spectra plots in POSEIDON.
+    
+    Args:
+        wl_min (float):
+            The minimum wavelength to plot.
+        wl_max (float):
+            The maximum wavelength to plot.
+        wl_axis (str, optional):
+            The type of x-axis to use ('log' or 'linear').
+            
+    Returns:
+        np.array:
+            The x axis tick values for the given wavelength range.
 
     '''
 
     wl_range = wl_max - wl_min
+
+    if (wl_max < wl_min):
+        raise Exception("Error: max wavelength must be greater than min wavelength.")
 
     # For plots over a wide wavelength range
[…]
