### 2022-12-22 1e40ffe
Minor plot function updates
