+++ POSEIDON/visuals.py
@@ -1033,7 +1033,7 @@ def set_spectrum_wl_ticks(wl_min, wl_max, wl_axis = 'log'):
         elif (wl_spacing == 3*np.power(10, major_exponent)):
             wl_spacing = 2*np.power(10, major_exponent)
 
-        wl_ticks = np.arange(round_sig_figs(wl_min, 2), round_sig_figs(wl_max, 2)+0.01, wl_spacing)
+        wl_ticks = np.arange(round_sig_figs(wl_min, 3), round_sig_figs(wl_max, 3)+0.0001, wl_spacing)
 
     return wl_ticks
 
@@ -1219,8 +1219,8 @@ def plot_spectra(spectra, planet, data_properties = None, show_data = False,
         # Loop over each model, finding the most extreme min / max range 
         for i in range(N_spectra):
             
-            transit_depth_min_i = np.min(spectra[i][0])
-            y_min_plt = min(y_min_plt, transit_depth_min_i)
+            y_min_i = np.min(spectra[i][0])
+            y_min_plt = min(y_min_plt, y_min_i)
             
         # Check if the lowest data point falls below the current y-limit
         if (show_data == True):
@@ -1240,8 +1240,8 @@ def plot_spectra(spectra, planet, data_properties = None, show_data = False,
         # Loop over each model, finding the most extreme min / max range 
         for i in range(N_spectra):
             
-            transit_depth_max_i = np.max(spectra[i][0])
-            y_max_plt = max(y_max_plt, transit_depth_max_i)
+            y_max_i = np.max(spectra[i][0])
+            y_max_plt = max(y_max_plt, y_max_i)
             
         # Check if the highest data point falls above the current y-limit
         if (show_data == True):
@@ -1712,8 +1712,7 @@ def plot_spectra_retrieved(spectra_median, spectra_low2, spectra_low1,
                            spectra_high1, spectra_high2, planet_name, 
                            data_properties, R_to_bin = 100, plt_label = None,
                            show_ymodel = True, wl_min = None, wl_max = None, 
-                           transit_depth_min = None, transit_depth_max = None,
-                           FpFs_min = None, FpFs_max = None, y_unit = 'transit_depth', 
+                           y_min = None, y_max = None, y_unit = 'transit_depth', 
                            colour_list = [], spectra_labels = [],
                            data_colour_list = [], data_labels = [],
                            data_marker_list = [], data_marker_size_list = [],
@@ -1724,6 +1723,16 @@ def plot_spectra_retrieved(spectra_median, spectra_low2, spectra_low1,
     
     '''
 
+    if (y_unit in ['(Rp/Rs)^2', '(Rp/R*)^2', 'transit_depth']):
+        plot_type = 'transmission'
+    elif (y_unit in ['Fp/Fs', 'Fp/F*', 'eclipse_depth']):
+        plot_type = 'emission'
+    elif (y_unit in ['Fp']):
+        plot_type = 'direct_emission'
+    else:
+        raise Exception("Unexpected y unit. Did you mean 'transit_depth' " +
+                       "or 'eclipse_depth'?")
+    
     # Find number of spectra to plot
     N_spectra = len(spectra_median)
 
@@ -1813,99 +1822,50 @@ def plot_spectra_retrieved(spectra_median, spectra_low2, spectra_low1,
             wl_max_i = np.max(spectra_median[i][1])
             wl_max = max(wl_max, wl_max_i)
 
-    # If the user did not specify a transit depth range, find min and max from input models
-    if (y_unit in ['(Rp/Rs)^2', '(Rp/R*)^2', 'transit_depth']):
-    
-        if (transit_depth_min == None):
[…]
