+++ POSEIDON/utility.py
@@ -1361,9 +1361,7 @@ def write_summary_file(results_prefix, planet_name, retrieval_name,
                 'Best reduced chi-square:\n',
                 '\n',
                 '-> chi^2_red = ' + stats_formatter.format(reduced_chi_square) + '\n',
-                '\n',
                 '-> degrees of freedom = ' + str(dof) + '\n',
-                '\n',
                 '-> chi^2 = ' + stats_formatter.format(chi_square) + '\n',
                 '\n',
                 '#################################\n']
@@ -1372,9 +1370,7 @@ def write_summary_file(results_prefix, planet_name, retrieval_name,
                 'Reduced chi-square undefined because N_params >= N_data!\n',
                 '\n',
                 '-> chi^2_red = Undefined\n',
-                '\n',
                 '-> degrees of freedom = Undefined\n',
-                '\n',
                 '-> chi^2 = ' + stats_formatter.format(chi_square) + '\n',
                 '\n',
                 '#################################\n']
