+++ POSEIDON/high_res_retrieval_demo.py
@@ -1,274 +1,0 @@
-# %% [markdown]
-# # Atmospheric Retrievals with POSEIDON
-# 
-# At long last, your proposal to observe the newly discovered hot Jupiter WASP-999b with the *Hubble Space Telescope* has been accepted. Congratulations! 
-# 
-# ### Loading Data
-# 
-# Months later, after carefully reducing the observations, you are ready to gaze in awe at your transmission spectrum.
-# 
-# First, you load all the usual stellar and planetary properties for this system.
-
-# %%
-from POSEIDON.core import create_star, create_planet
-from POSEIDON.constants import R_Sun, R_J, M_J
-
-#***** Define stellar properties *****#
-
-R_s = 1.21*R_Sun      # Stellar radius (m)
-T_s = 5605.0          # Stellar effective temperature (K)
-Met_s = -0.04         # Stellar metallicity [log10(Fe/H_star / Fe/H_solar)]
-log_g_s = 4.56        # Stellar log surface gravity (log10(cm/s^2) by convention)
-
-# Create the stellar object
-star = create_star(R_s, T_s, log_g_s, Met_s, stellar_spectrum = True, stellar_grid = 'phoenix')
-
-F_s = star['F_star']
-wl_s = star['wl_star']
-R_s = star['stellar_radius']
-
-
-#***** Define planet properties *****#
-
-planet_name = 'WASP-77Ab'  # Planet name used for plots, output files etc.
-
-R_p = 1.21*R_J      # Planetary radius (m)
-M_p = 0.07*M_J      # Mass of planet (kg)
-g_p = 4.3712        # Gravitational field of planet (m/s^2)
-T_eq = 1043.8       # Equilibrium temperature (K)
-
-# Create the planet object
-planet = create_planet(planet_name, R_p, mass = M_p, gravity = g_p, T_eq = T_eq)
-
-# If distance not specified, use fiducial value
-if (planet['system_distance'] is None):
-    planet['system_distance'] = 1    # This value only used for flux ratios, so it cancels
-d = planet['system_distance']
-
-# %% [markdown]
-# ### Creating a Retrieval Model
-# 
-# Now comes the creative part: what model do you try first to fit WASP-999b's transmission spectrum?
-# 
-# Given the a priori known low density of the planet, you conclude it is reasonable to assume this is a giant planet dominated by $\rm{H}_2$ and $\rm{He}$. Looking at your data above, especially the **huge** absorption feature in the infrared around 1.4 μm, you guess that $\rm{H}_2 \rm{O}$ could be present (based on theoretical predictions or after looking up its cross section).
-# 
-# So for a first attempt, you consider a model with $\rm{H}_2 \rm{O}$, an isothermal temperature profile, and no clouds.
-
-# %%
-from POSEIDON.core import define_model, wl_grid_constant_R
-from POSEIDON.utility import read_high_res_data
-#***** Define model *****#
-
-model_name = 'High-res retrieval'  # Model name used for plots, output files etc.
-
-bulk_species = ['H2', 'He']     # H2 + He comprises the bulk atmosphere
-param_species = ['H2O', 'CO']  # H2O, CO as in Brogi & Line
-
-# Create the model object
-model = define_model(model_name, bulk_species, param_species,
-                    PT_profile = 'Madhu', high_res=True)
-
-# Check the free parameters defining this model
-print("Free parameters: " + str(model['param_names']))
-
-#***** Wavelength grid *****#
-
-wl_min = 1.3      # Minimum wavelength (um)
-wl_max = 2.6      # Maximum wavelength (um)
[…]
