+++ POSEIDON/utility.py
@@ -380,19 +380,11 @@ def read_data(data_dir, fname, wl_unit = 'micron', bin_width = 'half',
     return wl_data, half_bin, spectrum, err
 
 
-<<<<<<< HEAD
 def read_high_res_data(data_dir, high_res = 'pca'): 
     '''
     Read an external dataset file. The expected file format is:
 
     TODO
-=======
-def read_high_res_data(data_dir): 
-    '''
-    Read an external dataset file. The expected file format is:
-
-    wavelength | bin half width | spectrum | error on spectrum
->>>>>>> 1e479b83e054c570b5ef6c53ccfdf72f4df0a2da
 
     Args:
         data_dir (str):
@@ -409,16 +401,11 @@ def read_high_res_data(data_dir):
     '''
     
     # Load data file
-<<<<<<< HEAD
     if high_res == 'pca':
-=======
-    try:
->>>>>>> 1e479b83e054c570b5ef6c53ccfdf72f4df0a2da
         wl_grid, data_arr = pickle.load(open(data_dir+'/PCA_matrix.pic', 'rb'))
         wl_grid, data_scale = pickle.load(open(data_dir+'/data_to_scale_with.pic', 'rb'))
         Phi = pickle.load(open(data_dir+'/ph.pic','rb'))                    # Time-resolved phases
         V_bary = pickle.load(open(data_dir+'/rvel.pic','rb'))               # Time-resolved Earth-star velocity (V_bary+V_sys) constructed in make_data_cube.py; then V_sys = V_sys_literature + d_V_sys
-<<<<<<< HEAD
         data = {
             'wl_grid': wl_grid, 
             'data_arr': data_arr,
@@ -513,92 +500,6 @@ def read_PT_file(PT_file_dir, PT_file_name, P_grid, P_unit = 'bar',
             Temperature profile from external file interpolated onto the
             POSEIDON model's pressure grid (K).
 
-=======
-    except:
-        print("Cannot locate data.")
-
-    data = {'wl_grid': wl_grid, 
-            'data_arr': data_arr,
-            'data_scale': data_scale,
-            'Phi': Phi,
-            'V_bary': V_bary }
-    return data
-    
-def read_spectrum(planet_name, fname, wl_unit = 'micron'):
-    '''
-    Read a previously computed spectrum from the POSEIDON output folder
-    (POSEIDON_output/planet_name/spectra).
-
-    Args:
-        planet_name (str):
-            Identifier for planet object (e.g. HD209458b).
-        fname (str):
-            Name of spectrum file.
-        wl_unit (str):
-            Unit of wavelength column (first column in file)
-            (Options: micron (or equivalent) / nm / A / m)
-
-    Returns:
-        wavelength (np.array of float): 
-            Model wavelength grid (μm).
-        spectrum (np.array of float):
-            Transmission or emission spectrum.
-
-    '''
-
-    # Load POSEIDON directory location where the spectrum is stored
-    input_dir = './POSEIDON_output/' + planet_name + '/spectra/'
-    
-    # Open file
-    data = pd.read_csv(input_dir + fname, sep = '[\s]{1,20}', 
-                       engine = 'python', header=None)
-
-    # Load wavelength then convert to μm
-    if (wl_unit in ['micron', 'um', 'μm']):
-        wavelength = np.array(data[0])
-    elif (wl_unit == 'nm'):
-        wavelength = np.array(data[0]) * 1e-3
-    elif (wl_unit == 'A'):
-        wavelength = np.array(data[0]) * 1e-4
-    elif (wl_unit == 'm'):
-        wavelength = np.array(data[0]) * 1e6
-    else:
[…]
