+++ POSEIDON/core.py
@@ -229,11 +229,7 @@ def define_model(model_name, bulk_species, param_species,
                  PT_dim = 1, X_dim = 1, cloud_dim = 1, TwoD_type = None, 
                  TwoD_param_scheme = 'difference', species_EM_gradient = [], 
                  species_DN_gradient = [], species_vert_gradient = [],
-<<<<<<< HEAD
                  surface = False, high_res = None, R_p_ref_enabled = True):
-=======
-                 surface = False, high_res = False):
->>>>>>> 1e479b83e054c570b5ef6c53ccfdf72f4df0a2da
     '''
     Create the model dictionary defining the configuration of the user-specified 
     forward model or retrieval.
@@ -310,14 +306,9 @@ def define_model(model_name, bulk_species, param_species,
             List of chemical species with a vertical mixing ratio gradient.
         surface (bool):
             If True, model a surface via an opaque cloud deck.
-<<<<<<< HEAD
         high_res (str):
             Define a model for high resolutional retrieval.
             (Options: 'pca', 'sysrem')
-=======
-        high_res (bool):
-            If True, define a model for high resolutional retrieval.
->>>>>>> 1e479b83e054c570b5ef6c53ccfdf72f4df0a2da
 
     Returns:
         model (dict):
@@ -380,11 +371,7 @@ def define_model(model_name, bulk_species, param_species,
                                       X_dim, cloud_dim, TwoD_type, TwoD_param_scheme, 
                                       species_EM_gradient, species_DN_gradient, 
                                       species_vert_gradient, Atmosphere_dimension,
-<<<<<<< HEAD
                                       opaque_Iceberg, surface, high_res, R_p_ref_enabled)
-=======
-                                      opaque_Iceberg, surface, high_res)
->>>>>>> 1e479b83e054c570b5ef6c53ccfdf72f4df0a2da
 
     # Package model properties
     model = {'model_name': model_name, 'object_type': object_type,
@@ -411,12 +398,8 @@ def define_model(model_name, bulk_species, param_species,
              'high_res_param_names': high_res_param_names,
              'N_params_cum': N_params_cum, 'TwoD_type': TwoD_type, 
              'TwoD_param_scheme': TwoD_param_scheme, 'PT_dim': PT_dim,
-<<<<<<< HEAD
              'X_dim': X_dim, 'cloud_dim': cloud_dim, 'surface': surface,
              'high_res': high_res
-=======
-             'X_dim': X_dim, 'cloud_dim': cloud_dim, 'surface': surface
->>>>>>> 1e479b83e054c570b5ef6c53ccfdf72f4df0a2da
             }
 
     return model
@@ -963,7 +946,6 @@ def compute_spectrum(planet, star, model, atmosphere, opac, wl,
 
         # Running POSEIDON on the CPU
         if (device == 'cpu'):
-<<<<<<< HEAD
         
             # Calculate extinction coefficients in standard mode
             kappa_clear, kappa_cloud = extinction(chemical_species, active_species,
@@ -1000,44 +982,6 @@ def compute_spectrum(planet, star, model, atmosphere, opac, wl,
             N_bf_species = len(bf_species)           # Number of bound-free species included
         
             # Calculate extinction coefficients in standard mode
-=======
-        
-            # Calculate extinction coefficients in standard mode
-            kappa_clear, kappa_cloud = extinction(chemical_species, active_species,
-                                                  CIA_pairs, ff_pairs, bf_species,
-                                                  n, T, P, wl, X, X_active, X_CIA, 
-                                                  X_ff, X_bf, a, gamma, P_cloud, 
-                                                  kappa_cloud_0, sigma_stored, 
-                                                  CIA_stored, Rayleigh_stored, 
[…]
