### 2022-12-24 68e6d1d
Added constant g hydrostatic solver (for tests)
