### 2022-12-27 5027c0a
add documentation for read_data
