### 2022-12-27 df7cd86
Add cross-correlation plot
