+++ POSEIDON/cross_correlation_WASP77.py
@@ -1,0 +1,164 @@
+from __future__ import absolute_import, unicode_literals, print_function
+from POSEIDON.high_res import cross_correlate_sysrem, fast_filter
+import math, os
+import numpy as np
+import pickle
+import pickle
+from scipy import constants
+from numba import jit
+from astropy.io import fits
+from scipy import interpolate
+from POSEIDON.core import create_star, create_planet, define_model, make_atmosphere, read_opacities, wl_grid_constant_R, wl_grid_line_by_line, compute_spectrum
+from POSEIDON.constants import R_Sun
+from POSEIDON.visuals import plot_stellar_flux
+from POSEIDON.constants import R_J, M_J
+import numpy as np
+from spectres import spectres
+from tqdm import tqdm
+from multiprocessing import Pool
+
+K_p = 192.06
+N_K_p = 200
+d_K_p = 1
+K_p_arr = (np.arange(N_K_p) - (N_K_p-1)//2) * d_K_p + K_p # making K_p_arr (centered on published or predicted K_p)
+# K_p_arr = [92.06 , ..., 191.06, 192.06, 193.06, ..., 291.06]
+
+V_sys = 0
+N_V_sys = 200
+d_V_sys = 1
+V_sys_arr = (np.arange(N_V_sys) - (N_V_sys-1)//2) * d_V_sys + V_sys # making V_sys_arr (centered on published or predicted V_sys (here 0 because we already added V_sys in V_bary))
+
+N_cores = 1 # Change how many cores to use here.
+core_indices = np.arange(N_cores)
+V_sys_arr_split = np.array_split(V_sys_arr, N_cores)
+
+def batch_cross_correlate(core_index, args):
+    F_s_obs, F_p_obs, wl = args
+    for i in range(core_index, len(V_sys_arr_split), N_cores):
+        print("Core "+str(i)+" is running.")
+        output_path = './CC_output/WASP-77Ab' # Could modify output path here.
+        data_path = './reference_data/observations/WASP-77Ab'
+        os.makedirs(output_path, exist_ok=True)
+
+        wl_grid, data_raw = pickle.load(open(data_path+'/data_RAW.pic', 'rb'))
+        Phi = pickle.load(open(data_path+'/ph.pic','rb'))                    # Time-resolved phases
+        V_bary = pickle.load(open(data_path+'/rvel.pic','rb'))               # Time-resolved Earth-star velocity (V_bary+V_sys) constructed in make_data_cube.py; then V_sys = V_sys_literature + d_V_sys
+        residuals, Us, Ws = fast_filter(data_raw, iter=15)
+        # We could also just use V_bary instead, then V_sys is just V_sys (not around zero anymore)
+        log_L_arr, CCF_arr = cross_correlate_sysrem(F_s_obs, F_p_obs, wl, K_p_arr, V_sys_arr_split[i], wl_grid, residuals, Us, V_bary, Phi)
+        pickle.dump([V_sys_arr_split[i], K_p_arr, CCF_arr, log_L_arr], open(output_path+'/test_sysrem_'+str(i)+'.pic','wb')) # N_cores of these produced. Will be read in by plot_CCF.py
+
+# The code below will only be run on one core to get the model spectrum.
+if __name__ == '__main__':
+    R_s = 1.21*R_Sun      # Stellar radius (m)
+    T_s = 5605.0          # Stellar effective temperature (K)
+    Met_s = -0.04         # Stellar metallicity [log10(Fe/H_star / Fe/H_solar)]
+    log_g_s = 4.56        # Stellar log surface gravity (log10(cm/s^2) by convention)
+
+    # Create the stellar object
[…]
