+++ POSEIDON/injection_test.py
@@ -1,0 +1,250 @@
+# %%
+from POSEIDON.core import create_star, create_planet
+from POSEIDON.constants import R_Sun, R_J, M_J
+
+#***** Define stellar properties *****#
+
+R_s = 0.91*R_Sun      # Stellar radius (m)
+T_s = 5605.0          # Stellar effective temperature (K)
+Met_s = -0.04         # Stellar metallicity [log10(Fe/H_star / Fe/H_solar)]
+log_g_s = 4.48        # Stellar log surface gravity (log10(cm/s^2) by convention)
+
+# Create the stellar object
+star = create_star(R_s, T_s, log_g_s, Met_s, stellar_spectrum = True, stellar_grid = 'phoenix')
+
+F_s = star['F_star']
+wl_s = star['wl_star']
+R_s = star['stellar_radius']
+
+#***** Define planet properties *****#
+
+planet_name = 'WASP-77Ab_injection_test'  # Planet name used for plots, output files etc.
+
+R_p = 1.21*R_J      # Planetary radius (m)
+M_p = 1.76*M_J      # Mass of planet (kg)
+g_p = 10**(3.4765-2)# Gravitational field of planet (m/s^2)
+T_eq = 1740         # Equilibrium temperature (K)
+
+# Create the planet object
+planet = create_planet(planet_name, R_p, mass = M_p, gravity = g_p, T_eq = T_eq)
+
+# If distance not specified, use fiducial value
+if (planet['system_distance'] is None):
+    planet['system_distance'] = 1    # This value only used for flux ratios, so it cancels
+d = planet['system_distance']
+
+planet['V_sin_i'] = 4.5
+
+# %%
+from POSEIDON.core import define_model, wl_grid_constant_R
+from POSEIDON.utility import read_high_res_data
+#***** Define model *****#
+
+model_name = 'High-res retrieval'  # Model name used for plots, output files etc.
+
+bulk_species = ['H2', 'He']     # H2 + He comprises the bulk atmosphere
+param_species = ['H2O', 'CO']  # H2O, CO as in Brogi & Line
+
+# Create the model object
+model = define_model(model_name, bulk_species, param_species,
+                    PT_profile = 'Madhu', high_res = 'pca', R_p_ref_enabled=False)
+
+# Check the free parameters defining this model
+print("Free parameters: " + str(model['param_names']))
+
+#***** Wavelength grid *****#
+
+wl_min = 1.3      # Minimum wavelength (um)
+wl_max = 2.6      # Maximum wavelength (um)
+R = 250000          # Spectral resolution of grid
+
+# wl = wl_grid_line_by_line(wl_min, wl_max)
+wl = wl_grid_constant_R(wl_min, wl_max, R)
+
+data_dir = './reference_data/observations/WASP-77Ab_injection_test'         # Special directory for this tutorial
+
+data = read_high_res_data(data_dir, high_res='pca')
+
+
+from POSEIDON.core import set_priors
+
+#***** Set priors for retrieval *****#
+
+# Initialise prior type dictionary
+prior_types = {}
+
+# Specify whether priors are linear, Gaussian, etc.
+prior_types['T_ref'] = 'uniform'
+prior_types['R_p_ref'] = 'uniform'
+prior_types['log_H2O'] = 'uniform'
+prior_types['log_CO'] = 'uniform'
+prior_types['log_CH4'] = 'uniform'
+prior_types['log_H2S'] = 'uniform'
+prior_types['log_NH3'] = 'uniform'
[…]
