### 2023-01-02 #4873 @ideasrule
@dfm Happy New Year!  I was waiting for the author to make the updates he said he was going to make ("community guidelines", rudimentary "automated tests", a bit more documentation on the visuals class).  Once those are done, I'm happy to put checkmarks on everything.  In general, I've found POSEIDON to be a very well written, useful, and easy to use code, so good job to @MartianColonist !
