### 2023-01-02 #4873 @LorenzoMugnai
Happy new year to everybody.
@dfm , I am in the same situation as @ideasrule. I'm ready to give my approval. I was just waiting for the author to come back with the mentioned updates. Other than that, I haven't found any reason to stop posting. POSEIDON is a very nice product and deserves this publication.
