### 2023-01-02 675b49a
Initial commit of automated testing
