+++ .github/pytest_testing.yml
@@ -1,0 +1,55 @@
+name: POSEIDON CI Pytest testing
+
+on:
+  push:
+    branches:    
+      - 'main'
+      - 'documentation'
+  pull_request:
+    branches:    
+      - 'main'
+      - 'documentation'
+
+jobs:
+  test:
+    name: Test on ${{ matrix.os }} with Python ${{ matrix.python-version }}
+    strategy:
+      matrix:
+        os: [ubuntu-latest, macos-latest]
+        python-version: [3.8, 3.9, 3.10]
+    steps:
+    - uses: actions/checkout@v3
+    - name: Set up Anaconda
+      uses: actions/setup-python@v3
+      with:
+        python-version: ${{ matrix.python-version }}
+        register: conda
+    - name: Create Anaconda environment
+      run: |
+        ${{ steps.conda.outputs.home }}/bin/conda create -n env python=${{ matrix.python-version }} -y
+        ${{ steps.conda.outputs.home }}/bin/conda activate env
+    - name: Install pymultinest
+      run: |
+        conda install -c conda-forge pymultinest -y
+    - name: Download input files
+      run: |
+        wget https://drive.google.com/file/d/1Rk_6sbIYC8c9La0fWHWpMPve6Jik7a3h/view?
+        usp=share_link -O input/opacity/Opacity_database_0.01cm-1.hdf5
+        wget https://drive.google.com/file/d/1hYLTzIy7cVicqGU8LHmLnq-3WQuyKISX/view?
+        usp=share_link -O input/opacity/Opacity_database_0.01cm-1_Temperate.hdf5
+        wget https://drive.google.com/file/d/1HA3gZUTmDIzZGFLTtuiPe6VDUxstxjZ_/view?
+        usp=share_link -O input/opacity/Opacity_database_cia.hdf5
+        wget https://drive.google.com/file/d/1xZzbVserwHZx0jmmhhEeQzk5RnxjFf2C/view?
+        usp=share_link -O input/stellar_grids.zip
+        unzip input/stellar_grids.zip -d input
+    - name: Set environment variables
+      run: |
+        echo 'export POSEIDON_input_data=input/opacity' >> $GITHUB_ENV
+        echo 'export PYSYN_CDBS=input/stellar_grids' >> $GITHUB_ENV
+    - name: Install package
+      run: |
+        python -m pip install --upgrade pip
+        pip install -e .
+    - name: Run tests
+      run: |
+        pytest
