### 2023-01-02 74fbd64
Fixed OS bug in testing CI
