+++ .github/workflows/pytest_testing.yml
@@ -17,7 +17,7 @@ jobs:
     strategy:
       matrix:
         os: [ubuntu-latest, macos-latest]
-        python-version: [3.8, 3.9, 3.10]
+        python-version: [3.9]
     steps:
     - uses: actions/checkout@v3
     - name: Set up Anaconda
