+++ .github/workflows/pytest_testing.yml
@@ -16,7 +16,7 @@ jobs:
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
-        os: [ubuntu-latest, macos-latest]
+        os: [ubuntu-latest]
         python-version: [3.9]
     steps:
     - uses: actions/checkout@v3
@@ -24,11 +24,10 @@ jobs:
       uses: actions/setup-python@v3
       with:
         python-version: ${{ matrix.python-version }}
-        register: conda
     - name: Create Anaconda environment
       run: |
-        ${{ steps.conda.outputs.home }}/bin/conda create -n env python=${{ matrix.python-version }} -y
-        ${{ steps.conda.outputs.home }}/bin/conda activate env
+        ${{ steps.setup-python.outputs.home }}/bin/conda create -n env python=${{ matrix.python-version }} -y
+        ${{ steps.setup-python.outputs.home }}/bin/conda activate env
     - name: Install pymultinest
       run: |
         conda install -c conda-forge pymultinest -y
