+++ .github/workflows/pytest_testing.yml
@@ -24,11 +24,15 @@ jobs:
       uses: actions/setup-python@v3
       with:
         python-version: ${{ matrix.python-version }}
+    - name: Add conda to PATH
+      run: |
+        echo 'export PATH=${{ steps.setup-python.outputs.home }}/bin:$PATH' >> $GITHUB_ENV
+    - name: Initialize conda
+      run: conda init bash
     - name: Create Anaconda environment
       run: |
-        ${{ steps.setup-python.outputs.home }}/bin/conda init bash
-        ${{ steps.setup-python.outputs.home }}/bin/conda create -n env python=${{ matrix.python-version }} -y
-        ${{ steps.setup-python.outputs.home }}/bin/conda activate env
+        conda create -n env python=${{ matrix.python-version }} -y
+        conda activate env
     - name: Install pymultinest
       run: |
         conda install -c conda-forge pymultinest -y
