+++ .github/workflows/pytest_testing.yml
@@ -27,14 +27,6 @@ jobs:
     - name: Add conda to PATH
       run: |
         echo 'export PATH=${{ steps.setup-python.outputs.home }}/bin:$PATH' >> $GITHUB_ENV
-    - name: Initialize conda
-      run: conda init bash
-    - name: Reopen shell
-      run: bash
-    - name: Create Anaconda environment
-      run: |
-        conda create -n env python=${{ matrix.python-version }} -y
-        conda activate env
     - name: Install pymultinest
       run: |
         conda install -c conda-forge pymultinest -y
