+++ .github/workflows/pytest_testing.yml
@@ -29,21 +29,7 @@ jobs:
         echo 'export PATH=${{ steps.setup-python.outputs.home }}/bin:$PATH' >> $GITHUB_ENV
     - name: Install pymultinest
       run: |
-        conda install -c conda-forge pymultinest -y
-    - name: Download input files
-      run: |
-        mkdir -p inputs/opacity
-        wget --load-cookies /tmp/cookies.txt "https://docs.google.com/uc?export=download&confirm=$(wget --quiet --save-cookies /tmp/cookies.txt --keep-session-cookies --no-check-certificate 'https://docs.google.com/uc?export=download&id=1xZzbVserwHZx0jmmhhEeQzk5RnxjFf2C' -O- | sed -rn 's/.*confirm=([0-9A-Za-z_]+).*/\1\n/p')&id=1xZzbVserwHZx0jmmhhEeQzk5RnxjFf2C" -O inputs/stellar_grids.zip && rm -rf /tmp/cookies.txt
-        wget --load-cookies /tmp/cookies.txt "https://docs.google.com/uc?export=download&confirm=$(wget --quiet --save-cookies /tmp/cookies.txt --keep-session-cookies --no-check-certificate 'https://docs.google.com/uc?export=download&id=1Rk_6sbIYC8c9La0fWHWpMPve6Jik7a3h' -O- | sed -rn 's/.*confirm=([0-9A-Za-z_]+).*/\1\n/p')&id=1Rk_6sbIYC8c9La0fWHWpMPve6Jik7a3h" -O inputs/opacity/Opacity_database_0.01cm-1.hdf5 && rm -rf /tmp/cookies.txt
-        wget --load-cookies /tmp/cookies.txt "https://docs.google.com/uc?export=download&confirm=$(wget --quiet --save-cookies /tmp/cookies.txt --keep-session-cookies --no-check-certificate 'https://docs.google.com/uc?export=download&id=1HA3gZUTmDIzZGFLTtuiPe6VDUxstxjZ_' -O- | sed -rn 's/.*confirm=([0-9A-Za-z_]+).*/\1\n/p')&id=1HA3gZUTmDIzZGFLTtuiPe6VDUxstxjZ_" -O inputs/opacity/Opacity_database_cia.hdf5 && rm -rf /tmp/cookies.txt
-        wget --load-cookies /tmp/cookies.txt "https://docs.google.com/uc?export=download&confirm=$(wget --quiet --save-cookies /tmp/cookies.txt --keep-session-cookies --no-check-certificate 'https://docs.google.com/uc?export=download&id=1hYLTzIy7cVicqGU8LHmLnq-3WQuyKISX' -O- | sed -rn 's/.*confirm=([0-9A-Za-z_]+).*/\1\n/p')&id=1hYLTzIy7cVicqGU8LHmLnq-3WQuyKISX" -O inputs/opacity/Opacity_database_0.01cm-1_Temperate.hdf5 && rm -rf /tmp/cookies.txt
-    - name: Extract input files
-      run: |
-        unzip inputs/stellar_grids.zip -d inputs
-    - name: Set environment variables
-      run: |
-        echo 'export POSEIDON_input_data=input/opacity' >> $GITHUB_ENV
-        echo 'export PYSYN_CDBS=input/stellar_grids' >> $GITHUB_ENV
+        conda install -c conda-forge mpi4py pymultinest -y
     - name: Install package
       run: |
         python -m pip install --upgrade pip
