+++ .github/workflows/pytest_testing.yml
@@ -27,9 +27,12 @@ jobs:
     - name: Add conda to PATH
       run: |
         echo 'export PATH=${{ steps.setup-python.outputs.home }}/bin:$PATH' >> $GITHUB_ENV
+    - name: Install mpi4py
+      run: |
+        conda install -c conda-forge mpi4py -y
     - name: Install pymultinest
       run: |
-        conda install -c conda-forge mpi4py pymultinest -y
+        conda install -c conda-forge pymultinest -y
     - name: Install package
       run: |
         python -m pip install --upgrade pip
