+++ .github/workflows/pytest_testing.yml
@@ -34,7 +34,7 @@ jobs:
     - name: Create conda environment
       run: |
         echo "begin: PATH=$PATH;"
-        conda create --name test_install python=3.9 -y
+        conda create --name test_install python=${{ matrix.python-version }} -y
         conda activate test_install || true
     - name: Install mpi4py
       run: |
@@ -49,6 +49,20 @@ jobs:
         conda activate test_install || true
         python -m pip install --upgrade pip
         pip install -e .
+    - name: Download input files
+      run: |
+        mkdir -p inputs/opacity
+        wget --load-cookies /tmp/cookies.txt "https://docs.google.com/uc?export=download&confirm=$(wget --quiet --save-cookies /tmp/cookies.txt --keep-session-cookies --no-check-certificate 'https://docs.google.com/uc?export=download&id=1xZzbVserwHZx0jmmhhEeQzk5RnxjFf2C' -O- | sed -rn 's/.*confirm=([0-9A-Za-z_]+).*/\1\n/p')&id=1xZzbVserwHZx0jmmhhEeQzk5RnxjFf2C" -O inputs/stellar_grids.zip && rm -rf /tmp/cookies.txt
+        wget --load-cookies /tmp/cookies.txt "https://docs.google.com/uc?export=download&confirm=$(wget --quiet --save-cookies /tmp/cookies.txt --keep-session-cookies --no-check-certificate 'https://docs.google.com/uc?export=download&id=1Rk_6sbIYC8c9La0fWHWpMPve6Jik7a3h' -O- | sed -rn 's/.*confirm=([0-9A-Za-z_]+).*/\1\n/p')&id=1Rk_6sbIYC8c9La0fWHWpMPve6Jik7a3h" -O inputs/opacity/Opacity_database_0.01cm-1.hdf5 && rm -rf /tmp/cookies.txt
+        wget --load-cookies /tmp/cookies.txt "https://docs.google.com/uc?export=download&confirm=$(wget --quiet --save-cookies /tmp/cookies.txt --keep-session-cookies --no-check-certificate 'https://docs.google.com/uc?export=download&id=1HA3gZUTmDIzZGFLTtuiPe6VDUxstxjZ_' -O- | sed -rn 's/.*confirm=([0-9A-Za-z_]+).*/\1\n/p')&id=1HA3gZUTmDIzZGFLTtuiPe6VDUxstxjZ_" -O inputs/opacity/Opacity_database_cia.hdf5 && rm -rf /tmp/cookies.txt
+        wget --load-cookies /tmp/cookies.txt "https://docs.google.com/uc?export=download&confirm=$(wget --quiet --save-cookies /tmp/cookies.txt --keep-session-cookies --no-check-certificate 'https://docs.google.com/uc?export=download&id=1hYLTzIy7cVicqGU8LHmLnq-3WQuyKISX' -O- | sed -rn 's/.*confirm=([0-9A-Za-z_]+).*/\1\n/p')&id=1hYLTzIy7cVicqGU8LHmLnq-3WQuyKISX" -O inputs/opacity/Opacity_database_0.01cm-1_Temperate.hdf5 && rm -rf /tmp/cookies.txt
+    - name: Extract input files
+      run: |
+        unzip inputs/stellar_grids.zip -d inputs
+    - name: Set environment variables
+      run: |
+        echo 'export POSEIDON_input_data=inputs/opacity' >> $GITHUB_ENV
+        echo 'export PYSYN_CDBS=inputs/stellar_grids' >> $GITHUB_ENV
     - name: Run tests
       run: |
         pytest
