+++ .github/workflows/pytest_testing.yml
@@ -20,7 +20,7 @@ jobs:
     strategy:
       matrix:
         os: [ubuntu-latest]
-        python-version: [3.9]
+        python-version: [3.9] 
     steps:
     - uses: actions/checkout@v3
     - name: Set up Anaconda
