+++ .github/workflows/pytest_testing.yml
@@ -23,6 +23,13 @@ jobs:
         python-version: [3.9] 
     steps:
     - uses: actions/checkout@v3
+    - name: free disk space
+      run: |
+        sudo swapoff -a
+        sudo rm -f /swapfile
+        sudo apt clean
+        docker rmi $(docker image ls -aq)
+        df -h
     - name: Set up Anaconda
       uses: conda-incubator/setup-miniconda@v2
       with:
