### 2023-01-03 e39c497
Free space to accommodate opacity files for tests
