### 2023-01-03 5adbbf7
Restored large opacity database for next test
