+++ .github/workflows/pytest_testing.yml
@@ -110,6 +110,7 @@ jobs:
         wget --load-cookies /tmp/cookies.txt "https://docs.google.com/uc?export=download&confirm=$(wget --quiet --save-cookies /tmp/cookies.txt --keep-session-cookies --no-check-certificate 'https://docs.google.com/uc?export=download&id=1JPWlorWm48WKqI_XPs7OKZ3_0_8VyoXc' -O- | sed -rn 's/.*confirm=([0-9A-Za-z_]+).*/\1\n/p')&id=1JPWlorWm48WKqI_XPs7OKZ3_0_8VyoXc" -O inputs/stellar_grids.zip && rm -rf /tmp/cookies.txt
         wget --load-cookies /tmp/cookies.txt "https://docs.google.com/uc?export=download&confirm=$(wget --quiet --save-cookies /tmp/cookies.txt --keep-session-cookies --no-check-certificate 'https://docs.google.com/uc?export=download&id=1_jGwMyfhutVZxC0OSklcVnsvPck_sCtj' -O- | sed -rn 's/.*confirm=([0-9A-Za-z_]+).*/\1\n/p')&id=1_jGwMyfhutVZxC0OSklcVnsvPck_sCtj" -O inputs/opacity/Opacity_database_cia.hdf5 && rm -rf /tmp/cookies.txt
         wget --load-cookies /tmp/cookies.txt "https://docs.google.com/uc?export=download&confirm=$(wget --quiet --save-cookies /tmp/cookies.txt --keep-session-cookies --no-check-certificate 'https://docs.google.com/uc?export=download&id=1ajoOsCRN8DaUNHVd3KrC2rtA7BkNtUrc' -O- | sed -rn 's/.*confirm=([0-9A-Za-z_]+).*/\1\n/p')&id=1ajoOsCRN8DaUNHVd3KrC2rtA7BkNtUrc" -O inputs/opacity/Opacity_database_0.01cm-1_Temperate.hdf5 && rm -rf /tmp/cookies.txt
+        wget --load-cookies /tmp/cookies.txt "https://docs.google.com/uc?export=download&confirm=$(wget --quiet --save-cookies /tmp/cookies.txt --keep-session-cookies --no-check-certificate 'https://docs.google.com/uc?export=download&id=12WnIKCJbJmvW-tJURa_kXBeAL4EtZ-OX' -O- | sed -rn 's/.*confirm=([0-9A-Za-z_]+).*/\1\n/p')&id=12WnIKCJbJmvW-tJURa_kXBeAL4EtZ-OX" -O inputs/opacity/Opacity_database_0.01cm-1.hdf5 && rm -rf /tmp/cookies.txt
 
     - name: Cache downloaded files
       if: steps.restore_cache.outputs.cache-hit != 'true'
@@ -130,6 +131,7 @@ jobs:
       run: |
         echo 'export POSEIDON_input_data=inputs/opacity' >> $GITHUB_ENV
         echo 'export PYSYN_CDBS=inputs/stellar_grids' >> $GITHUB_ENV
+        source ~/.bashrc
 
     - name: Run tests
       run: |
