### 2023-01-03 #4873 @MartianColonist
A very happy new year to you all!

I have almost finished the requested updates (thanks for the great suggestions @ideasrule and @LorenzoMugnai!). I'll push the new version in the next day or so.

The one stumbling block I've encountered is with the automated tests. Since POSEIDON requires quite large input files to be downloaded (40+GB, dominated by the cross section database), GitHub Actions CI is proving problematic. The package can install correctly, but the automated download of the cross section database fails because Google Drive consistently blocks wget partway through the transfer of the largest file. Moving the files to Zenodo didn't help either, since downloads from Zenodo are sufficiently slow that GitHub Actions times out. I'd certainly appreciate any suggestions!

If there isn't an obvious way to get CI working, would it be ok to have the tests in the /tests folder without automation (such that a user can run them after installing the package)?
