### 2023-01-03 #4873 @dfm
@MartianColonist — thanks for the update! Automation isn't strictly required for JOSS, but always good to have if you can. One option is to have a subset of tests that don't require the full download + some offline tests that aren't integrated into the CI system. But as I said this isn't formally required as far as I'm concerned.
