### 2023-01-03 #4873 @ideasrule
> The one stumbling block I've encountered is with the automated tests. Since POSEIDON requires quite large input files to be downloaded (40+GB, dominated by the cross section database), GitHub Actions CI is proving problematic. The package can install correctly, but the automated download of the cross section database fails because Google Drive consistently blocks wget partway through the transfer of the largest file. Moving the files to Zenodo didn't help either, since downloads from Zenodo are sufficiently slow that GitHub Actions times out. I'd certainly appreciate any suggestions!

Here's a suggestion: if you don't have tests that depend on the true cross sections, and I'm guessing you don't, what about making up the cross section files on the fly using a random number generator?  Of course, it would be better for the order of magnitude to be approximately correct, not off by 10^10.  This has the added advantage that if you make the cross sections the same at all altitudes (but not at all wavelengths), you can analytically calculate the transit spectrum and compare it to what your code outputs.
