### 2023-01-04 6712dff
Added docstrings for visuals.py
