+++ POSEIDON/visuals.py
@@ -3,7 +3,6 @@ Plotting routines to visualise POSEIDON output.
 
 '''
 
-from enum import unique
 import os
 import numpy as np
 import scipy.constants as sc
@@ -88,7 +87,42 @@ def scale_lightness(colour_name, scale):
 def plot_transit(ax, R_p, r, T, phi, phi_edge, dphi, theta, theta_edge, dtheta, 
                  perspective, plot_labels = True):
     '''
-    
+    Subfunction used by the 'plot_geometry' function below. This function plots
+    a 2D slice through an exoplanet and its atmosphere (to scale) from various 
+    observer perspectives.
+
+    Args:
+        ax (matplotlib axis object):
+            A matplotlib axis instance.
+        R_p (float):
+            White light planetary radius.
+        r (3D np.array of float):
+            Radial distance profile (m).
+        T (3D np.array of float):
+            Temperature profile (K).
+        phi (np.array of float):
+            Mid-sector angles (radians).
+        phi_edge (np.array of float):
+            Boundary angles for each sector (radians).
+        dphi (np.array of float):
+            Angular width of each sector (radians).
+        theta (np.array of float):
+            Mid-zone angles (radians).
+        theta_edge (np.array of float):
+            Boundary angles for each zone (radians).
+        dtheta (np.array of float):
+            Angular width of each zone (radians).
+        perspective (str):
+            Observer viewing perspective for 2D slice.
+            (Options: terminator / day-night).
+        plot_labels (bool)
+            If False, removes text labels from the plot.
+
+    Returns:
+        p (matplotlib PatchCollection):
+            Patch collection containing atmosphere slices and temperature 
+            colour bar. The figure itself is written to the provided axis.
+
     '''
 
     ax.axis('equal')
@@ -316,26 +350,26 @@ def plot_geometry(planet, star, model, atmosphere, plot_labels = True):
     '''
     Plots two 2D slice plots through the planetary atmosphere (to scale),
     coloured according to the temperature field. The left panel corresponds
-    to a slice through the terminator plane, whilst the right panel is a
+    to a slice through the terminator plane, while the right panel is a
     slice through the north pole - observer plane.
 
     Args:
-        planet (dict of str: various):
-            Collection of planetary properties used by POSEIDON.
-        star (dict of str: various):
-            Collection of stellar properties used by POSEIDON.
-        model (dict of str: various):
-            A specific description of a given POSEIDON model.
-        atmosphere (dict of str: various):
-            Collection of atmospheric properties.
+        planet (dict):
+            POSEIDON planet properties dictionary.
+        star (dict):
+            POSEIDON stellar properties dictionary (currently unused).
+        model (dict):
+            POSEIDON model properties dictionary.
+        atmosphere (dict):
+            POSEIDON atmospheric properties dictionary.
         plot_labels (bool)
             If False, removes text labels from the plot.
 
     Returns:
-        fig (matplotlib figure object)
+        fig (matplotlib figure object):
+            The geometric slice plot.
 
     '''
-
 
     # Unpack model and atmospheric properties
     planet_name = planet['planet_name']
@@ -391,30 +425,49 @@ def plot_geometry(planet, star, model, atmosphere, plot_labels = True):
 def plot_PT(planet, model, atmosphere, show_profiles = [],
             PT_label = None, log_P_min = None, log_P_max = None, T_min = None,
             T_max = None, colour = 'darkblue', legend_location = 'lower left'):
-    ''' 
+    '''
     Plot the pressure-temperature (P-T) profiles defining the atmosphere.
     
-    For a 1D model, a single P-T profile is plotted. For 2D or 3D models,
[…]
