### 2023-01-04 4cf1210
Updated paper and docs to mention emission spectra
