+++ joss/paper.md
@@ -11,13 +11,15 @@ tags:
 authors:
   - name: Ryan J. MacDonald
     orcid: 0000-0003-4816-3469
-    affiliation: "1, 2"
+    affiliation: "1, 2, 3"
 affiliations:
   - name: Department of Astronomy, University of Michigan, 1085 S. University Ave., Ann Arbor, MI 48109, USA
     index: 1
-  - name: Department of Astronomy and Carl Sagan Institute, Cornell University, 122 Sciences Drive, Ithaca, NY 14853, USA
+  - name: NHFP Sagan Fellow
     index: 2
-date: 04 November 2022
+  - name: Department of Astronomy and Carl Sagan Institute, Cornell University, 122 Sciences Drive, Ithaca, NY 14853, USA
+    index: 3
+date: 04 January 2023
 bibliography: paper.bib
 
 aas-doi: 10.3847/1538-4357/ac47fe
@@ -33,7 +35,7 @@ Exoplanet atmospheres are a dynamic and fast-changing field at the frontier of m
 
 # Exoplanet Modelling and Atmospheric Retrieval with `POSEIDON`
 
-The first major use case for `POSEIDON` is 'forward modelling' --- illustrated on the left of \autoref{fig:POSEIDON_architecture}. A user can generate a model planet spectrum, for a given star-planet system, by providing a specific set of atmospheric properties (e.g. the chemical composition and temperature). The forward model mode allows users to explore how atmospheric properties alter an exoplanet spectrum and to produce predicted model spectra for observing proposals. The required input files (pre-computed stellar grids and an opacity database) are available to download from a Zenodo repository (linked in the documentation).
+The first major use case for `POSEIDON` is 'forward modelling' --- illustrated on the left of \autoref{fig:POSEIDON_architecture}. A user can generate a model planet spectrum, for a given star-planet system, by providing a specific set of atmospheric properties (e.g. the chemical composition and temperature). The forward model mode allows users to explore how atmospheric properties alter an exoplanet spectrum and to produce predicted model spectra for observing proposals. The required input files (pre-computed stellar grids and an opacity database) are available to download from an online repository (linked in the documentation).
 
 The second major use case for `POSEIDON` is atmospheric retrieval --- illustrated on the right of \autoref{fig:POSEIDON_architecture}. To initialise a retrieval, a user provides an observed exoplanet spectrum and the range of atmospheric properties to be explored (i.e. the prior ranges for a set of free parameters defining a model). A Bayesian statistical sampling algorithm --- nominally [`PyMultiNest`](https://github.com/JohannesBuchner/PyMultiNest) [@Buchner:2014] --- then repeatedly calls the forward model, comparing the generated spectrum to the observations, until the parameter space is fully explored and a convergence criteria reached. The main outputs of an atmospheric retrieval are the posterior probability distributions of the model parameters and the model's Bayesian evidence. The Bayesian evidences from multiple retrievals, in turn, can be subsequently compared to compute detection significances for each model component (e.g. the statistical confidence for a molecule being present in the planetary atmosphere).
 
@@ -45,9 +47,13 @@ The second major use case for `POSEIDON` is atmospheric retrieval --- illustrate
 
 Recent years have seen a substantial improvement in the number of high-quality exoplanet spectra. In particular, the newly operational JWST and a profusion of high-resolution ground-based spectrographs offer an abundance of exoplanet data. The accurate interpretation of such data requires a retrieval code that can rapidly explore complex parameter spaces describing a rich variety of atmospheric phenomena.
 
[…]
