+++ docs/index.rst
@@ -8,9 +8,9 @@ from exoplanet spectra. POSEIDON has two main components: (1) a 'forward' model,
 generates 1D, 2D, or 3D transmission spectra; and (2) a nested sampling retrieval 
 framework that uses the sampling algorithm `PyMultiNest <https://johannesbuchner.github.io/PyMultiNest/>`_,
 wrapped around TRIDENT, to explore the range of atmospheric properties consistent
-with an observed exoplanet spectrum.
+with an observed exoplanet transmission spectrum.
 
-POSEIDON's features currently include:
+POSEIDON's official features currently include:
 
 * Transmission spectra modelling for 1D, 2D, and 3D exoplanet atmospheres.
 * Rapid atmospheric retrievals that can run on your laptop.
@@ -18,13 +18,21 @@ POSEIDON's features currently include:
 * Parametric prescriptions for stellar contamination, multidimensional clouds, and more.
 * High-resolution line-by-line models (:math:`R \sim 10^6`) for cross correlation analyses.
 
-The initial public release of POSEIDON contains tutorials on generating model 
-transmission spectra with TRIDENT and running atmospheric retrievals.
+Beta features:
+
+* Emission spectra modelling and retrievals for 1D, cloud-free atmospheres without scattering.
+
+The initial public release of POSEIDON contains a range of tutorials on 
+generating forward models and a tutorial on running atmospheric retrievals.
 Tutorials on multidimensional retrievals will be added soon.
 
-POSEIDON is available under the BSD 3-Clause License. If you use POSEIDON for retrievals, 
-please cite `MacDonald & Madhusudhan (2017) <https://ui.adsabs.harvard.edu/abs/2017MNRAS.469.1979M/abstract>`_. 
-If you only use the forward model, TRIDENT, please cite `MacDonald & Lewis (2022) 
+POSEIDON is available under the BSD 3-Clause License. If you use POSEIDON,
+please cite `MacDonald & Madhusudhan (2017) 
+<https://ui.adsabs.harvard.edu/abs/2017MNRAS.469.1979M/abstract>`_ and 
+`MacDonald (2023) <https://joss.theoj.org/papers/69710c0498d02fd1c6a0cfa4b01af7c5>`_.
+Additionally, if you make use of the multidimensional transmission spectra 
+modelling capabilities, we would appreciate a citation for the TRIDENT methods
+paper: `MacDonald & Lewis (2022) 
 <https://ui.adsabs.harvard.edu/abs/2021arXiv211105862M/abstract>`_.
 
 .. toctree::
