+++ docs/conf.py
@@ -18,12 +18,12 @@ sys.path.insert(0, os.path.abspath('../'))
 # -- Project information -----------------------------------------------------
 
 project = 'POSEIDON'
-copyright = '2022, Ryan J. MacDonald'
+copyright = '2023, Ryan J. MacDonald'
 author = 'Ryan J. MacDonald'
 master_doc = 'index'
 
 # The full version, including alpha/beta/rc tags
-release = '2022'
+release = '2023'
 
 
 # -- General configuration ---------------------------------------------------
