+++ POSEIDON/absorption.py
@@ -142,14 +142,14 @@ def T_interpolation_init(N_T_fine, T_grid, T_fine, y):
     
     w_T = np.zeros(N_T_fine)
     
-    # Find T index in cross secion arrays prior to fine temperature value
+    # Find T index in cross section arrays prior to fine temperature value
     for j in range(N_T_fine):
         
-        if (T_fine[j] < T_grid[0]):   # If fine temperature point falls off LHS of temperaure grid
+        if (T_fine[j] < T_grid[0]):   # If fine temperature point falls off LHS of temperature grid
             y[j] = -1                 # Special value (-1) stored, interpreted in interpellator
             w_T[j] = 0.0              # Weight not used in this case
             
-        elif (T_fine[j] >= T_grid[-1]):   # If fine temperature point falls off RHS of temperaure grid
+        elif (T_fine[j] >= T_grid[-1]):   # If fine temperature point falls off RHS of temperature grid
             y[j] = -2                     # Special value (-2) stored, interpreted in interpellator
             w_T[j] = 0.0                  # Weight not used in this case
         
@@ -660,7 +660,7 @@ def shared_memory_array(rank, comm, shape):
 
 def opacity_tables(rank, comm, wl_model, chemical_species, active_species, 
                    cia_pairs, ff_pairs, bf_species, T_fine, log_P_fine,
-                   opacity_database = 'High-T'):
+                   opacity_database = 'High-T', testing = False):
     ''' 
     Initialisation function to read in and pre-interpolate all opacities.
         
@@ -733,18 +733,27 @@ def opacity_tables(rank, comm, wl_model, chemical_species, active_species,
                             "your .bashrc or .bash_profile to point to the " +
                             "directory containing the POSEIDON opacity database.")
         
-        # Open HDF5 files containing molecular + atomic opacities
-        if (opacity_database == 'High-T'):        # High T database
-            opac_file = h5py.File(opacity_path + 'Opacity_database_0.01cm-1.hdf5', 'r')  
-        elif (opacity_database == 'Temperate'):   # Low T database
-            opac_file = h5py.File(opacity_path + 'Opacity_database_0.01cm-1_Temperate.hdf5', 'r')
+        # If running the automated GitHub tests, don't read the opacity database
+        if (testing == True):
+            opac_file = ''
+            log_P_grid = np.array([-6, -5, -4, -3, -2, -1, 0, 1, 2])
+            N_P = len(log_P_grid)
+
+        # For all other applications
+        else:
+
+            # Open HDF5 files containing molecular + atomic opacities
+            if (opacity_database == 'High-T'):        # High T database
+                opac_file = h5py.File(opacity_path + 'Opacity_database_0.01cm-1.hdf5', 'r')  
+            elif (opacity_database == 'Temperate'):   # Low T database
+                opac_file = h5py.File(opacity_path + 'Opacity_database_0.01cm-1_Temperate.hdf5', 'r')
+
+            # Read P grid used in opacity files
+            log_P_grid = np.array(opac_file['H2O/log(P)'])   # Units: log10(P/bar) - H2O choice arbitrary, all P grids are the same
+            N_P = len(log_P_grid)                            # No. of pressures in opacity files
         
         # Open HDF5 files containing collision-induced absorption (CIA)
         cia_file = h5py.File(opacity_path + 'Opacity_database_cia.hdf5', 'r')
-        
-        #***** Read in P grid used in opacity files*****#
-        log_P_grid = np.array(opac_file['H2O/log(P)'])   # Units: log10(P/bar) - H2O choice arbitrary, all P grids are the same
-        N_P = len(log_P_grid)                            # No. of pressures in opacity files
         
[…]
