### 2023-01-04 a20c38d
Updated test path, added Python 3.10
