### 2023-01-04 a1c79c7
Increment Python version to test 3.10
