### 2023-01-04 28cd719
JOSS paper action only triggers on updating main
