+++ .github/workflows/draft-pdf.yml
@@ -1,5 +1,9 @@
 name: "Build JOSS Paper Draft"
-on: [push]
+
+on:
+  push:
+    branches:    
+      - 'main'
 
 jobs:
   paper:
