+++ .github/workflows/pytest_testing.yml
@@ -41,14 +41,54 @@ jobs:
         auto-update-conda: true
         auto-activate-base: true
 
+    - name: Create conda environment
+      run: |
+        echo "begin: PATH=$PATH;"
+        conda create --name test_install python=${{ matrix.python-version }} -y
+
+    - name: Install mpi4py and pymultinest
+      run: |
+        conda activate test_install || true
+        conda install -c conda-forge mpi4py -y
+        conda install -c conda-forge pymultinest -y
+
+    - name: Install POSEIDON
+      run: |
+        conda activate test_install || true
+        python -m pip install --upgrade pip
+        pip install -e .
+
+    - name: Restore cached input files (if already downloaded)
+      if: steps.cache.outputs.cache-hit == 'true'
+      uses: actions/cache@v2
+      with:
+        path: |
+          inputs/opacity
+          inputs/stellar_grids
+        key: ${{ runner.os }}-${{ hashFiles('inputs/opacity/*') }}-${{ hashFiles('inputs/stellar_grids/*') }}
+
+    - name: Download input files (only CIA opacity for CI tests)
+      if: steps.restore_cache.outputs.cache-hit != 'true'
+      run: |
+        mkdir -p inputs/opacity
+        wget --load-cookies /tmp/cookies.txt "https://docs.google.com/uc?export=download&confirm=$(wget --quiet --save-cookies /tmp/cookies.txt --keep-session-cookies --no-check-certificate 'https://docs.google.com/uc?export=download&id=1HA3gZUTmDIzZGFLTtuiPe6VDUxstxjZ_' -O- | sed -rn 's/.*confirm=([0-9A-Za-z_]+).*/\1\n/p')&id=1HA3gZUTmDIzZGFLTtuiPe6VDUxstxjZ_" -O inputs/opacity/Opacity_database_cia.hdf5 && rm -rf /tmp/cookies.txt
+
+    - name: Cache downloaded files
+      if: steps.restore_cache.outputs.cache-hit != 'true'
+      uses: actions/cache@v2
+      with:
+        path: |
+          inputs/opacity
+        key: ${{ runner.os }}-${{ hashFiles('inputs/opacity/*') }}-${{ hashFiles('inputs/stellar_grids/*') }}
+        restore-keys: |
+          ${{ runner.os }}-
+
     - name: Set environment variables
       run: |
-
-        echo "::debug::$GITHUB_WORKSPACE"
-        echo "::debug::format('{0}/{1}', '$GITHUB_WORKSPACE', 'inputs/opacity/')"
-
-        echo "POSEIDON_input_data=format('{0}/{1}', '$GITHUB_WORKSPACE', 'inputs/opacity/')" >> $GITHUB_ENV
-        echo "PYSYN_CDBS='inputs/stellar_grids/'" >> $GITHUB_ENV
-
+        echo "POSEIDON_input_data='$HOME/inputs/opacity/'" >> $GITHUB_ENV
         echo "::debug::$GITHUB_ENV"
 
+    - name: Run tests
+      run: |
+        conda activate test_install || true
+        pytest
