+++ POSEIDON/utility.py
@@ -331,7 +331,7 @@ def read_data(data_dir, fname, wl_unit = 'micron', bin_width = 'half',
     '''
     
     # Load data file
-    data = pd.read_csv(data_dir + '/' + fname, sep = '[\s]{1,20}', 
+    data = pd.read_csv(data_dir + '/' + fname, sep = '[\\s]{1,20}', 
                        header = None, skiprows = skiprows, engine = 'python')
 
     # Load wavelength and half bin width, then convert both to μm
@@ -405,7 +405,7 @@ def read_spectrum(planet_name, fname, wl_unit = 'micron'):
     input_dir = './POSEIDON_output/' + planet_name + '/spectra/'
     
     # Open file
-    data = pd.read_csv(input_dir + fname, sep = '[\s]{1,20}', 
+    data = pd.read_csv(input_dir + fname, sep = '[\\s]{1,20}', 
                        engine = 'python', header=None)
 
     # Load wavelength then convert to μm
@@ -460,7 +460,7 @@ def read_PT_file(PT_file_dir, PT_file_name, P_grid, P_unit = 'bar',
         PT_file_dir = os.path.abspath(os.path.join(os.path.dirname( __file__ ), 
                                       '.', 'reference_data/models/TRAPPIST-1e/'))
     
-    PT_file = pd.read_csv(PT_file_dir + '/' + PT_file_name, sep = '[\s]{1,20}', 
+    PT_file = pd.read_csv(PT_file_dir + '/' + PT_file_name, sep = '[\\s]{1,20}', 
                           header = None, skiprows = skiprows, engine = 'python')
     
     # Read pressure
@@ -534,7 +534,7 @@ def read_chem_file(chem_file_dir, chem_file_name, P_grid, chem_species_in_file,
         chem_file_dir = os.path.abspath(os.path.join(os.path.dirname( __file__ ), 
                                         '.', 'reference_data/models/TRAPPIST-1e/'))
 
-    chem_file_input = pd.read_csv(chem_file_dir + '/' + chem_file_name, sep = '[\s]{1,20}', 
+    chem_file_input = pd.read_csv(chem_file_dir + '/' + chem_file_name, sep = '[\\s]{1,20}', 
                                   header = None, skiprows = skiprows, engine = 'python')
     
     # Read pressure and mixing ratios
@@ -761,7 +761,7 @@ def read_retrieved_spectrum(planet_name, model_name, retrieval_name = None):
     fname = output_dir + retrieval_name + '_spectrum_retrieved.txt'
 
     # Read retrieved spectrum confidence intervals
-    spec_file = pd.read_csv(fname, sep = '[\s]{1,20}', engine = 'python', 
+    spec_file = pd.read_csv(fname, sep = '[\\s]{1,20}', engine = 'python', 
                             header = None, skiprows = 1)
 
     wl = np.array(spec_file[0])           # Wavelengths (um)
@@ -791,7 +791,7 @@ def read_retrieved_PT(planet_name, model_name, retrieval_name = None):
     fname = output_dir + retrieval_name + '_PT_retrieved.txt'
 
     # Read retrieved temperature confidence intervals
-    PT_file = pd.read_csv(fname, sep = '[\s]{1,20}', engine = 'python', 
+    PT_file = pd.read_csv(fname, sep = '[\\s]{1,20}', engine = 'python', 
                           header = None, skiprows = 1)
 
     P = np.array(PT_file[0])         # Pressure (bar)
@@ -856,7 +856,7 @@ def read_retrieved_log_X(planet_name, model_name, retrieval_name = None):
     for q in range(N_species):
 
         # Read retrieved mixing ratio confidence intervals
[…]
