### 2023-01-04 7b89f5e
Switch to workspace directory in env variable
