+++ .github/workflows/pytest_testing.yml
@@ -85,8 +85,7 @@ jobs:
 
     - name: Set environment variables
       run: |
-        echo "POSEIDON_input_data=$HOME/inputs/opacity/" >> $GITHUB_ENV
-        echo "::debug::$GITHUB_ENV"
+        echo "POSEIDON_input_data=$GITHUB_WORKSPACE/inputs/opacity/" >> $GITHUB_ENV
 
     - name: Run tests
       run: |
