### 2023-01-04 c3fc190
Restored stellar grids for CI test
