### 2023-01-04 57266de
Pytest CI testing operational
