+++ .github/workflows/pytest_testing.yml
@@ -1,4 +1,4 @@
-name: POSEIDON CI Pytest testing
+name: Pytest testing
 
 on:
   push:
@@ -73,4 +73,4 @@ jobs:
     - name: Run tests
       run: |
         conda activate test_install || true
-        pytest
+        pytest -rA
