### 2023-01-05 6a68cf4
Improved installation instructions format
