+++ docs/content/installation.rst
@@ -14,81 +14,48 @@ Before installing POSEIDON, you will need to download these input files
 * `Opacity_database_0.01cm-1_Temperate.hdf5 <https://drive.google.com/file/d/1hYLTzIy7cVicqGU8LHmLnq-3WQuyKISX/view?usp=sharing>`_
 * `Stellar_grids.zip <https://drive.google.com/file/d/1xZzbVserwHZx0jmmhhEeQzk5RnxjFf2C/view?usp=sharing>`_
 
-Place the three opacity files in a directory called 'opacity' in a location of 
-your choice and unzip the stellar grid folder. I usually place these in an
-'input' folder with the structure 'input/opacity' and 'input/stellar_grids'.
-
-.. note::
-   For Windows users, we recommend installing `Windows Subsystem for Linux (WSL) 
-   <https://docs.microsoft.com/en-us/windows/wsl/about>`_
-   before proceeding. WSL provides a Linux environment you can use on Windows.
-   
-   One of the main modules used by POSEIDON, PyMultiNest, does not natively 
-   support Windows. However, it will work fine if you use WSL to install and 
-   run POSEIDON.
-
-Next, you need to create two new environment variables: one pointing to the 
-directory containing the opacity files and a second to the (unzipped) stellar
-grid folder. 
-
-Linux environment variables
----------------------------
-  
-Enter the following lines into a terminal:
-
-.. code-block:: bash
-
-   echo 'export POSEIDON_input_data="/𝗣𝗔𝗧𝗛/𝗧𝗢/𝗬𝗢𝗨𝗥/𝗜𝗡𝗣𝗨𝗧/𝗗𝗜𝗥𝗘𝗖𝗧𝗢𝗥𝗬/opacity/"' >>~/.bashrc
-   echo 'export PYSYN_CDBS="/𝗣𝗔𝗧𝗛/𝗧𝗢/𝗬𝗢𝗨𝗥/𝗜𝗡𝗣𝗨𝗧/𝗗𝗜𝗥𝗘𝗖𝗧𝗢𝗥𝗬/stellar_grids/"' >>~/.bashrc
-
-You should replace the bold text above with the location of your input directory.
-
-Alternatively, you can just open your .bashrc file (a hidden file in your home 
-directory) in a text editor and add the following two lines at the bottom:
-
-.. code-block:: bash
-
-   export POSEIDON_input_data="/𝗣𝗔𝗧𝗛/𝗧𝗢/𝗬𝗢𝗨𝗥/𝗜𝗡𝗣𝗨𝗧/𝗗𝗜𝗥𝗘𝗖𝗧𝗢𝗥𝗬/opacity/"
-   export PYSYN_CDBS="/𝗣𝗔𝗧𝗛/𝗧𝗢/𝗬𝗢𝗨𝗥/𝗜𝗡𝗣𝗨𝗧/𝗗𝗜𝗥𝗘𝗖𝗧𝗢𝗥𝗬/stellar_grids/"
-
-Mac OS environment variables
-----------------------------
-
-Check whether your terminals use bash (older OS) or zsh (newer OS).
-
-For bash terminals, enter the following lines:
-   
-.. code-block:: bash
-
[…]
