+++ README.rst
@@ -7,6 +7,22 @@
    </div>
    <br/><br/>
 
+.. image:: https://readthedocs.org/projects/poseidon-retrievals/badge/?version=latest
+   :target: https://poseidon-retrievals.readthedocs.io/en/latest/?badge=latest
+   :alt: Documentation Status
+
+.. image:: https://img.shields.io/badge/License-BSD_3--Clause-blue.svg
+   :target: https://github.com/MartianColonist/POSEIDON/blob/main/LICENSE
+   :alt: GitHub License
+
+.. image:: https://github.com/MartianColonist/POSEIDON/actions/workflows/pytest_testing.yml/badge.svg
+   :target: https://github.com/MartianColonist/POSEIDON/actions/workflows/pytest_testing.yml
+   :alt: Pytest Test Suite Status
+
+.. image:: https://joss.theoj.org/papers/69710c0498d02fd1c6a0cfa4b01af7c5/status.svg
+   :target: https://joss.theoj.org/papers/69710c0498d02fd1c6a0cfa4b01af7c5
+   :alt: Journal of Open Source Software Paper Status
+
 **A Python Package for Multidimensional Atmospheric Retrievals of Exoplanet Spectra**
 
 POSEIDON is a spectral modelling and atmospheric retrieval code for exoplanets.
@@ -40,27 +56,14 @@ POSEIDON is available under the BSD 3-Clause License.
 Please see the LICENSE file for more information.
 
 Contributing to POSEIDON
--------
+------------------------
 
-POSEIDON is currently in the open beta stage of testing, so any feedback, bug reports,
-or questions are certainly welcome! Please reach out to ryanjmac@umich.edu
+For bug reports or feature requests, please raise an issue.
 
+Interesting in contributing to POSEIDON? Please see the `contribution guide 
+<https://poseidon-retrievals.readthedocs.io/en/latest/content/contributing.html>`_.
 
-.. image:: https://readthedocs.org/projects/poseidon-retrievals/badge/?version=latest
-   :target: https://poseidon-retrievals.readthedocs.io/en/latest/?badge=latest
-   :alt: Documentation Status
+Questions and Support
+---------------------
 
-.. image:: https://img.shields.io/badge/License-BSD_3--Clause-blue.svg
-   :target: https://github.com/MartianColonist/POSEIDON/blob/main/LICENSE
-   :alt: GitHub License
-
-.. image:: https://github.com/MartianColonist/POSEIDON/actions/workflows/pytest_testing.yml/badge.svg
-   :target: https://github.com/MartianColonist/POSEIDON/actions/workflows/pytest_testing.yml
-   :alt: Pytest Test Suite Status
-
-.. image:: https://joss.theoj.org/papers/69710c0498d02fd1c6a0cfa4b01af7c5/status.svg
-   :target: https://joss.theoj.org/papers/69710c0498d02fd1c6a0cfa4b01af7c5
-   :alt: Journal of Open Source Software Paper Status
-
-
-
+Feel free to reach out to ryanjmac@umich.edu
