### 2023-01-05 b58b081
Add tests for macOS
