+++ tests/test_retrieval.py
@@ -1,0 +1,195 @@
+import pytest
+
+
+def test_continuum_retrieval():
+    '''
+    Test POSEIDON's retrieval functionality on a synthetic WASP-121b 
+    transmission spectrum only including H-, H2-H2 and H2-He CIA, and Rayleigh
+    scattering (i.e. a continuum-only retrieval).
+    
+    '''
+
+    from POSEIDON.constants import R_Sun, R_J
+    from POSEIDON.core import create_star, create_planet, define_model, \
+                            make_atmosphere, read_opacities, wl_grid_constant_R, \
+                            compute_spectrum, load_data, set_priors
+    from POSEIDON.instrument import generate_syn_data_from_user
+    from POSEIDON.visuals import plot_spectra, plot_spectra_retrieved
+    from POSEIDON.utility import plot_collection, read_retrieved_spectrum
+    from POSEIDON.retrieval import run_retrieval
+    from POSEIDON.corner import generate_cornerplot
+    import scipy.constants as sc
+    import pymultinest
+    import numpy as np
+    import os
+
+    #***** Define stellar properties *****#
+
+    R_s = 1.46*R_Sun      # Stellar radius (m)
+    T_s = 6776.0          # Stellar effective temperature (K)
+    Met_s = 0.13          # Stellar metallicity [log10(Fe/H_star / Fe/H_solar)]
+    log_g_s = 4.24        # Stellar log surface gravity (log10(cm/s^2) by convention)
+
+    # Create the stellar object
+    star = create_star(R_s, T_s, log_g_s, Met_s)
+
+    #***** Define planet properties *****#
+
+    planet_name = 'WASP-121b'  # Planet name used for plots, output files etc.
+
+    R_p = 1.753*R_J      # Planetary radius (m)
+    log_g_p = 2.97       # Gravitational field of planet (cgs)
+    T_eq = 2450          # Equilibrium temperature (K)
+
+    # Create the planet object
+    planet = create_planet(planet_name, R_p, log_g = log_g_p, T_eq = T_eq)
+
+    #***** Model wavelength grid *****#
+
+    wl_min = 0.4    # Minimum wavelength (um)
+    wl_max = 1.8    # Maximum wavelength (um)
+    R = 1000   # We can get away with R = 1k for this test, since we only have continuum opacity
+
+    # We need to provide a model wavelength grid to initialise instrument properties
+    wl = wl_grid_constant_R(wl_min, wl_max, R)
+
+    #***** Define model *****#
+
+    model_name = 'H-_retrieval_test'
+
+    bulk_species = ['H2', 'He']
+    param_species = ['H-']
+
+    # Create the model object
+    model = define_model(model_name, bulk_species, param_species,
+                        PT_profile = 'isotherm')
+
+    # Specify the pressure grid of the atmosphere
+    P_min = 1.0e-7    # 0.1 ubar
+    P_max = 100       # 100 bar
+    N_layers = 100    # 100 layers
+
+    # We'll space the layers uniformly in log-pressure
+    P = np.logspace(np.log10(P_max), np.log10(P_min), N_layers)
+
+    # Specify the reference pressure and radius
+    P_ref = 10.0   # Reference pressure (bar)
+    R_p_ref = R_p   # Radius at reference pressure
+
+    # Provide a specific set of model parameters for the atmosphere 
+    PT_params = np.array([T_eq])         
+    log_X_params = np.array([-9.0])
+
+    # Generate the atmosphere
+    atmosphere = make_atmosphere(planet, model, P, P_ref, R_p_ref, 
+                                PT_params, log_X_params)
+
+    #***** Read opacity data *****#
+
+    opacity_treatment = 'opacity_sampling'
[…]
