+++ docs/rtd_requirements.txt
@@ -1,4 +1,4 @@
 nbsphinx
 sphinx-autoapi
-sphinx_rtd_theme>=0.3.1
+sphinx_rtd_theme=1.0.0
 ipython
