### 2023-01-05 #4873 @MartianColonist
The new version is live!

Major changes:

1. New 'Contributing to POSEIDON' page in the documentation.
2. Two automated tests via GitHub Action (see below): (i) numerical transmission spectrum vs. analytic solution; (ii) atmospheric retrieval of simulated data without Gaussian scatter for a planet with H- opacity (i.e. continuum opacity only).
3. Emission spectra beta support is now mentioned in the documentation and paper.
4. Docstrings added for visuals.py (I also fixed the wavelength ticks issue @ideasrule mentioned).

Regarding the analytic test, here is POSEIDON compared to the analytic formula from Betremieux & Swain (2017):

![Example Planet_Analytic Test_transmission_spectra](https://user-images.githubusercontent.com/22718554/210701177-2929ed48-1d86-489b-aa75-cbb9575f6a65.png)

The agreement is within 0.01 % (< 1 ppm).

The retrieval test takes about 1 minute to run and successfully recovers the true values of each parameter to < 0.5 %:

![WASP-121b_Continuum Retrieval Test_retrieved_spectra](https://user-images.githubusercontent.com/22718554/210701184-2f97cbbd-2a30-4d70-a0b1-3924af6b2f9f.png)

![H-_retrieval_test_corner](https://user-images.githubusercontent.com/22718554/210701072-2226f569-1b89-4f6c-b7d2-25b56f55ec33.png)

Thanks for your very helpful input @LorenzoMugnai, @ideasrule, and @dfm!
