### 2023-01-06 #4873 @ideasrule
The changes look great, @MartianColonist !   I've checked off every item in the checklist.  I only have three very minor suggestions left:

Your code's biggest selling point is that it supports 2D and 3D exoplanet atmospheres.  I think you should mention this on the GitHub README.

For PLATON, the most up to date reference is Zhang et al 2020 (https://ui.adsabs.harvard.edu/abs/2020ApJ...899...27Z/abstract).

When I ran the references check, it reported that "Ariel: Enabling planetary science across light-years" is missing a DOI.  You could use the arXiv DOI: https://doi.org/10.48550/arXiv.2104.04824

Great job developing POSEIDON!
