+++ joss/paper.bib
@@ -53,6 +53,25 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@ARTICLE{Zhang:2020,
+       author = {{Zhang}, Michael and {Chachan}, Yayaati and {Kempton}, Eliza M. -R. and {Knutson}, Heather A. and {Chang}, Wenjun (Happy)},
+        title = "{PLATON II: New Capabilities and a Comprehensive Retrieval on HD 189733b Transit and Eclipse Data}",
+      journal = {Astrophysical Journal},
+     keywords = {Exoplanet atmospheres, Exoplanet atmospheric composition, Astronomical models, Transmission spectroscopy, 487, 2021, 86, 2133, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics},
+         year = 2020,
+        month = aug,
+       volume = {899},
+       number = {1},
+          eid = {27},
+        pages = {27},
+          doi = {10.3847/1538-4357/aba1e6},
+archivePrefix = {arXiv},
+       eprint = {2004.09513},
+ primaryClass = {astro-ph.EP},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2020ApJ...899...27Z},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
 @ARTICLE{Molliere:2019,
        author = {{Molli{\`e}re}, P. and {Wardenier}, J.~P. and {van Boekel}, R. and {Henning}, Th. and {Molaverdikhani}, K. and {Snellen}, I.~A.~G.},
         title = "{petitRADTRANS. A Python radiative transfer package for exoplanet characterization and retrieval}",
@@ -468,6 +487,7 @@ archivePrefix = {arXiv},
         month = nov,
           eid = {arXiv:2104.04824},
         pages = {arXiv:2104.04824},
+          doi = {10.48550/arXiv.2104.04824},
 archivePrefix = {arXiv},
        eprint = {2104.04824},
  primaryClass = {astro-ph.IM},
