### 2023-01-10 #4873 @dfm
@MartianColonist — Things are looking good! I've opened a small PR with some minor edits to the manuscript, please take a look and merge or let me know what you think. It would also be great if you can address [the final suggestions](https://github.com/openjournals/joss-reviews/issues/4873#issuecomment-1372984035) from @ideasrule above.

Once you've done that:
