### 2023-01-12 4ee938e
Added acknowledgements to the JOSS superstars
