### 2023-01-12 #4873 @MartianColonist
@dfm, I've implemented the suggestions from @ideasrule and merged your pull request.

1. The manuscript looks good!
2. The version is 1.0.0.
3. The archive doi is: 10.5281/zenodo.7530543
