### 2023-01-13 #4873 @dfm
@ideasrule, @LorenzoMugnai — Many thanks for your reviews here! JOSS relies upon the volunteer effort of people like you and we simply wouldn't be able to do this without you!!
